extract sections of Nonmem control streams

This is a very commonly used wrapper for the input part of the model file. Look NMextractText for more general functionality suitable for the results part too.

NMgetSection(
  file = NULL,
  lines = NULL,
  text = NULL,
  section,
  return = "text",
  keepEmpty = FALSE,
  keepName = TRUE,
  keepComments = TRUE,
  asOne = TRUE,
  simplify = TRUE,
  cleanSpaces = FALSE,
  ...
)

Arguments

file
A file to read from. Normally a .mod or .lst. See lines also.

lines
Text lines to process. This is an alternative to using the file argument.

text
Use this argument if the text to process is one long character string, and indicate the line seperator with the linesep argument (handled by NMextractText). Use only one of file, lines, and text.

section
The name of section to extract. Examples: "INPUT", "PK", "TABLE", etc.

return
If "text", plain text lines are returned. If "idx", matching line numbers are returned. "text" is default.

keepEmpty
Keep empty lines in output? Default is FALSE.

keepName
Keep the section name in output (say, "$PROBLEM") Default is TRUE. It can only be FALSE, if return"idx".

keepComments
Keep comment lines?

asOne
If multiple hits, concatenate into one. This will most often be relevant with name="TABLE". If FALSE, a list will be returned, each element representing a table. Default is TRUE. So if you want to process the tables separately, you probably want FALSE here.

simplify
If asOne=FALSE, do you want the result to be simplified if only one table is found? Default is TRUE which is desirable for interactive analysis. For programming, you probably want FALSE.

cleanSpaces
If TRUE, leading and trailing are removed, and multiplied succeeding white spaces are reduced to single white spaces.

...
Additional arguments passed to NMextractText

Value

character vector with extracted lines.

See also

Examples

NMgetSection(NMdata_filepath("examples/nonmem/run001.lst"),section="DATA")

#> [1] "$DATA      PK.csv IGNORE=@ IGNORE=(PKFL.EQ.0)"      
#> [2] "            IGNORE=(TRTPN.EQ.1) IGNORE=(TRTPN.EQ.3)"
#> [3] "            IGNORE=(TRTPN.EQ.4)"