Read input data formatted for Nonmem

This function is especially useful if the csv file was written using NMwriteData.

NMreadCsv(file, args.fread, as.fun = NULL)

Arguments

file

The file to read. Must be pure text.

args.fread

List of arguments passed to fread. Notice that except for "file", you need to supply all arguments to fread if you use this argument. Default values can be configured using NMdataConf.

as.fun

The default is to return data as a data.frame. Pass a function (say tibble::as_tibble) in as.fun to convert to something else. If data.tables are wanted, use as.fun="data.table". The default can be configured using NMdataConf.

Value

A data set of class as defined by as.fun.

Details

This is almost just a shortcut to fread so you don't have to remember how to read the data that was exported for Nonmem. The only added feature is that meta data as written by NMwriteData is read and attached as NMdata metadata before data is returned.

See also