Read comments to parameter definitions in Nonmem control streams

When interpreting parameter estimates, it is often needed to recover information about the meaning of the different parameters from control stream. `NMreadParsText` provides a flexible way to organize the comments in the parameter sections into a `data.frame`. This can subsequently easily be merged with parameter values as obtained with `NMreadExt`.

NMreadParsText(
  file,
  fields,
  fields.omega = fields,
  fields.sigma = fields.omega,
  use.theta.nums = FALSE,
  spaces.split = FALSE,
  modelname,
  col.model,
  as.fun
)

Arguments

file

Path to the control stream to read.

fields

Defines naming and splitting of contents of lines in parameter sections. Default is "%init;%symbol;%num;%label;%unit".

fields.omega

Like `fields`, applied to `$OMEGA` section. Default is to reuse `fields`.

fields.sigma

Like `fields`, applied to `$SIGMA` section. Default is to reuse `fields.omega`.

use.theta.nums

If the num field should be used to number thetas. I do not see where this is advantageous to do.

spaces.split

Is a blank in `fields` to be treated as a field seperator? Default is not to (i.e. neglect spaces in `fields`).

modelname

See ?NMscanData

col.model

See ?NMscanData

as.fun

See ?NMscanData

Details

Off-diagonal omega and sigma elements will only be correctly treated if their num field specifies say 1-2 to specify it is covariance between 1 and 2.