read all output data tables in nonmem run

read all output data tables in nonmem run

NMscanTables(file, details = F, as.fun, quiet, tab.count = TRUE)

Arguments

file
the nonmem file to read (normally .mod or .lst)

details
If TRUE, metadata is added to output. In this case, you get a list. Typically, this is mostly useful if programming up functions which behavior must depend on properties of the output.

as.fun
The default is to return data as a data.frame. Pass a function (say tibble::as_tibble) in as.fun to convert to something else. If data.tables are wanted, use as.fun="data.table". The default can be configured using NMdataConf.

quiet
The default is to give some information along the way on what data is found. But consider setting this to TRUE for non-interactive use. Default can be configured using NMdataConf.

tab.count
Nonmem includes a counter of tables in the written data files. These are often not useful. However, if tab.count is TRUE (default), this will be carried forward and added as a column called TABLENO.

Value

A list of all the tables as data.frames. If details=TRUE, this is in one element, called data, and meta is another element. If not, only the element corresponding to data is returned.

See also