Write dataset for use in Nonmem (and R)

Instead of trying to remember the arguments to pass to write.csv, use this wrapper. It tells you what to write in $DATA and $INPUT in nonmem, and it (additionally) exports an rds or Rdata file as well which is highly preferable for use in R. It never edits the data before writing the datafile. The filenames for csv, rds etc. are derived by replacing the extension to the filename given in the file argument.

NMwriteData(
  data,
  file,
  write.csv = TRUE,
  write.RData = FALSE,
  write.rds = write.csv,
  script,
  args.stamp,
  args.rds,
  nmdrop,
  nmdir.data,
  col.flag = "FLAG"
)

Arguments

data
The dataset to write to file for use in Nonmem.

file
The file to write to. The extension (everything after and including last ".") is dropped. csv, rds and other standard file name extensions are added.

write.csv
Write to csv file?

write.RData
In case you want to save to .RData object. Not recommended. Use write.rds instead.

write.rds
write an rds file?

script
If provided, the object will be stamped with this script name before saved to rds or Rdata. See ?stampObj.

args.stamp
A list of arguments to be passed to stampObj.

args.rds
A list of arguments to be passed to saveRDS.

nmdrop
Only used for generation of proposed text for Nonmem control stream. Columns to drop in Nonmem $DATA. This has two implications. One is that the proposed $DATA indicates =DROP after the given column names. The other that in case it is a non-numeric column, succeeding columns can still be included.

nmdir.data
For the $DATA text proposal only. The path to the input datafile to be used in the Nonmem $DATA section. Often, a relative path to the actual Nonmem run is wanted here.

col.flag
Name of a numeric column with zero value for rows to include in Nonmem run, non-zero for rows to skip. The argument is only used for generating the proposed text to paste into the Nonmem control stream. To skip this feature, use col.flag=NULL.

Value

Text for inclusion in Nonmem control stream, invisibly.

Details

When writing csv files, the file will be comma-separated. Because Nonmem does not support quoted fields, you must avoid commas in character fields. An error is returned if commas are found in strings.

The user is provided with text to use in Nonmem. This lists names of the data columns. Once a column is reached that Nonmem will not be able to read as a numeric and column is not in nmdrop, the list is stopped.

See also