Create character vectors without quotation marks

When creating character vectors with several elements, it becomes a lot of quotes to type. cc provides a simple way to skip the quotes - but only for simple strings.

cc(...)

Arguments

...

The unquoted names that will become character values in the returned vector.

Details

Don't use cc with any special characters - only alphanumerics and no spaces supported.

See also

Examples

cc(a,b,`a b`)
#> [1] "a"   "b"   "a b"
cc(a,b,"a b")
#> [1] "a"   "b"   "a b"
## be careful with spaces and special characters
cc( d)
#> [1] "d"
cc(" d")
#> [1] " d"
cc()
#> character(0)