Assign exclusion flags to a dataset based on specified table

The aim with this function is to take a (say PK) dataset and a pre-specified table of flags, assign the flags automatically.

flagsAssign(
  data,
  tab.flags,
  subset.data,
  col.flagn,
  col.flagc,
  flags.increasing = FALSE,
  as.fun = NULL
)

Arguments

data
The dataset to assign flags to.

tab.flags
A data.frame containing at least these named columns: FLAG, flag, condition. Condition is disregarded for FLAG==0. FLAG must be numeric and non-negative, flag and condition are characters.

subset.data
An optional string that provides a subset of data to assign flags to. A common example is subset=\"EVID==0\" to only assign to observations. Numerical and character flags will be missing in rows that are not matched by this subset.

col.flagn
The name of the column containing the numerical flag values in tab.flags. This will be added to data. Default value is FLAG and can be configured using NMdataConf.

col.flagc
The name of the column containing the character flag values in tab.flags. This will be added to data. Default value is flag and can be configured using NMdataConf.

flags.increasing
The flags are applied by either decreasing (default) or increasing value of col.flagn. By using decreasing order, you can easily adjust the Nonmem IGNORE statement from IGNORE(FLAG.NE.0) to say IGNORE(FLAG.GT.10) if BLQ's have FLAG=10, and you decide to include these in the analysis.

as.fun
The default is to return data.tables if input data is a data.table, and return a data.frame for all other input classes. Pass a function in as.fun to convert to something else. If return.all=F, this is applied to data and tab.flags independently.

Value

The dataset with flags added. See parameter flags.return as well.

Details

dt.flags must contain a column with numerical exclusion flags, one with character exclusion flags, and one with a expressions to evaluate for whether to apply the exclusion flag. The flags are applied sequentially, by increasing value of the numerical exclusion flag.

See also

Examples

pk <- readRDS(file=system.file("examples/data/xgxr2.rds",package="NMdata"))
dt.flags <- data.frame(
       flagn=10,
       flagc="Below LLOQ",
       condition=c("BLQ==1")
)
pk <- flagsAssign(pk,dt.flags,col.flagn="flagn",col.flagc="flagc")

#> Coding flagn = 10, flagc = Below LLOQ

unique(pk[,c("flagn","flagc","BLQ")])

#>    flagn        flagc BLQ
#> 1:     0 Analysis set   0
#> 2:    10   Below LLOQ   1

flagsCount(pk[EVID==0],dt.flags,col.flagn="flagn",col.flagc="flagc")

#>                 flagc N.left Nobs.left N.discard N.disc.cum Nobs.discard
#> 1: All available data    150      1352        NA          0           NA
#> 2:         Below LLOQ    131       755        19         19          597
#> 3:       Analysis set    131       755        NA         19           NA
#>    Nobs.disc.cum
#> 1:             0
#> 2:           597
#> 3:           597