Create an overview of number of retained and discarded datapoints.

Generate an overview of number of observations disregarded due to different reasons. And how many are left after each exclusion flag.

flagsCount(
  data,
  tab.flags,
  file,
  col.id = "ID",
  col.flagn,
  col.flagc,
  by = NULL,
  flags.increasing = FALSE,
  flagc.0 = "Analysis set",
  name.all.data = "All available data",
  grp.incomp = "EVID",
  save = TRUE,
  as.fun = NULL
)

Arguments

data

The dataset including both FLAG and flag columns.

tab.flags

A data.frame containing at least these named columns: FLAG, flag, condition. Condition is disregarded for FLAG==0.

file

A file to write the table of flag counts to. Will probably be removed and put in a separate function.

col.id

The name of the subject ID column. Default is "ID".

col.flagn

The name of the column containing the numerical flag values in tab.flags. This will be added to data. Use the same as when flagsAssign was called (if that was used). Default value is FLAG and can be configured using NMdataConf.

col.flagc

The name of the column containing the character flag values in data and tab.flags. Use the same as when flagsAssign was called (if that was used). Default value is flag and can be configured using NMdataConf.

by

An optional column to group the counting by. This could be "STUDY", "DRUG", "EVID", or a combination of multiple columns.

flags.increasing

The flags are applied by either decreasing (default) or increasing value of col.flagn. By using decreasing order, you can easily adjust the Nonmem IGNORE statement from IGNORE(FLAG.NE.0) to say IGNORE(FLAG.GT.10) if BLQ's have FLAG=10, and you decide to include these in the analysis.

flagc.0

The character flag to assign to rows that are not matched by exclusion conditions (numerical flag 0).

name.all.data

What to call the total set of data before applying exclusion flags. Default is "All available data".

grp.incomp

Column(s) that distinct incompatible subsets of data. Default is "EVID" meaning that if different values of EVID are found in data, the function will return an error. This is a safeguard not to mix data unintentionally when counting flags.

save

Save file? Default is TRUE, meaning that a file will be written if file argument is supplied.

as.fun

The default is to return a data.table if input data is a data.table, and return a data.frame for all other input classes. Pass a function in as.fun to convert to something else. If data is not a data.table, default can be configured using NMdataConf.

Value

A summary table with number of discarded and retained subjects and observations when applying each condition in the flag table. "discarded" means that the reduction of number of observations and subjects resulting from the flag, "retained" means the numbers that are left after application of the flag. The default is "both" which will report both. Class as defined by as.fun.

Details

This function is used to count flags as assigned by the flagsAssign function.

Notice that the character flags reported in the output table are taken from tab.flags. The data column named by the value of col.flagc (default is flag) is not used.

In the returned table, N.discarded is the difference in number of subjects since previous step. If two is reported, it can mean that the remaining one observation of these two subjects are discarded due to this flag. The majority of the samples can have been discarded by earlier flags.

See also

Examples

pk <- readRDS(file=system.file("examples/data/xgxr2.rds",package="NMdata"))
dt.flags <- data.frame(
       flagn=10,
       flagc="Below LLOQ",
       condition=c("BLQ==1")
)
pk <- flagsAssign(pk,dt.flags,subset.data="EVID==0",col.flagn="flagn",col.flagc="flagc")
#> Coding flagn = 10, flagc = Below LLOQ
pk <- flagsAssign(pk,subset.data="EVID==1",flagc.0="Dosing",
        col.flagn="flagn",col.flagc="flagc")
unique(pk[,c("EVID","flagn","flagc","BLQ")])
#>    EVID flagn        flagc BLQ
#> 1:    1     0       Dosing   0
#> 2:    0    10   Below LLOQ   1
#> 3:    0     0 Analysis set   0
flagsCount(pk[EVID==0],dt.flags,col.flagn="flagn",col.flagc="flagc")
#>                 flagc N.left Nobs.left N.discard N.disc.cum Nobs.discard
#> 1: All available data    150      1352        NA          0           NA
#> 2:         Below LLOQ    131       755        19         19          597
#> 3:       Analysis set    131       755        NA         19           NA
#>    Nobs.disc.cum
#> 1:             0
#> 2:           597
#> 3:           597