Pretty wrapping of lines in NMdata vignettes

Pretty wrapping of lines in NMdata vignettes

messageWrap(
  ...,
  fun.msg = message,
  prefix = "\n",
  initial = "",
  width,
  track.msg = FALSE
)

Arguments

...

parameters to pass to strwrap

fun.msg

The function to pass the text through. Typically, message, warning, or stop. If NULL, nothing will happen, and NULL is invisibly returned.

prefix

Passed to strwrap. Default is "\n".

initial

Passed to strwrap. Default is an empty string.

width

Passed to strwrap. Default is 80.

track.msg

If TRUE, the name of the function throwing the message/warning/error is mentioned. This is not default but useful when using function inside other functions.

Value

Nothing.