Determine as.fun to use based on argument and options

Determine as.fun to use based on argument and options

runAsFun(data, as.fun)

Arguments

data
Dataset to possibly convert.

as.fun
A function to apply if not NULL.

Value

Possibly converted data.

Details

NMdata functions can return objects based on different classes. The general default is to return data.frames. However, if you prefer working with with other classes (typically data.table or tibble), you can have those returned instead. In general this is controlled by the run.as arguments to functions and the "NMdata.as.fun" option.

You can use the following values for run.as (argument) and NMdata.as.fun (option):

"none"Return a data.table. Yes, a data.table and not a data.frame. If you work with data.table, this will give the best performance.

a functionApply the function. You must supply the actual function, not a string representing it. Examples: as.data.frame or tibble::as_tibble. Do not use as.data.table - use "none" to get a data.table.

NULLUse default, i.e. try getOptions("NMdata.as.fun") and if still NULL, rely on default behavior.

It is perfectly possible but not recommended to convert the results coming out of NMdata functions for three reasons

ConvenienceYou would lose the NMdata class which enables a method like summary to work properly.

ReliabilityYou would have to do it every time you have called an NMdata function increasing risk of errors.

Performancecan be significantly better using "none".

The functions that return table data structures (like a data.frame) take an argument as.fun which controls this behaviour. You can also control the default behaviour by the NMdata.as.fun option. If not NULL (the default value) as.fun always overrules getOption("NMdata.run.as").

For functions that read and process data from the file system (like NMreadCsv, NMscanData, NMscanTables and many others), default is data.frame. Default can be overruled using options(NMdata.run.as=...), and the default behaviour can be overruled using the run.as argument.

For functions that process a data set supplied in an argument (as opposed to reading data from the file system), the class of the input data makes a difference. This applies to functions like mergeCheck, NMorderCols, findCovs, findVars, flagsAssign, flagsCount. If a data.table is supplied to these (and as.fun is NULL), a data.table will be returned. If not, a data.frame is still default, and the default can be configured using options(NMdata.run.as=...).

NMdata heavily uses data.table under the hood, so "none" means that no conversion is done by the end of the function. Since NMdata has no other dependencies than data.table, it cannot convert to other classes unless the user supply the conversion function.

Examples

dat <- NMscanData(NMdata_filepath("examples/nonmem/xgxr001.lst"))

#> In NMscanData(NMdata_filepath("examples/nonmem/xgxr001.lst")) : Input data
#> columns will be appended to output data. However, column(s) were identified as
#> unique identifiers, present in both input and output data. If this column or
#> one of these columns is not modified by the Nonmem run, consider using this in
#> col.row for a robust merge of input and output data. Candidate columns: ROW To
#> skip this check, please specify either col.row (recommended) or
#> cbind.by.filters.

#> Model:  xgxr001 
#> 
#> Tables, number of columns in tables, and their detail level:
#>            table used/total level
#>  xgxr001_res.txt      16/16   row
#>            input      22/22   row
#> 
#> Numbers of ID's and rows in data
#>       N From output tables From input data only
#>  N.rows                905                    0
#>   N.ids                150                    0
#> 
#> Distribution of rows on event types
#>  EVID From output tables
#>     0                755
#>     1                150

class(dat)

#> [1] "NMdata"     "data.frame"

dat <- NMscanData(NMdata_filepath("examples/nonmem/xgxr001.lst"),as.fun="none")

#> In NMscanData(NMdata_filepath("examples/nonmem/xgxr001.lst"), as.fun = "none")
#> : Input data columns will be appended to output data. However, column(s) were
#> identified as unique identifiers, present in both input and output data. If
#> this column or one of these columns is not modified by the Nonmem run,
#> consider using this in col.row for a robust merge of input and output data.
#> Candidate columns: ROW To skip this check, please specify either col.row
#> (recommended) or cbind.by.filters.

#> Model:  xgxr001 
#> 
#> Tables, number of columns in tables, and their detail level:
#>            table used/total level
#>  xgxr001_res.txt      16/16   row
#>            input      22/22   row
#> 
#> Numbers of ID's and rows in data
#>       N From output tables From input data only
#>  N.rows                905                    0
#>   N.ids                150                    0
#> 
#> Distribution of rows on event types
#>  EVID From output tables
#>     0                755
#>     1                150

class(dat)

#> [1] "NMdata"     "data.table" "data.frame"

options(NMdata.run.as="none")
dat <- NMscanData(NMdata_filepath("examples/nonmem/xgxr001.lst"))

#> In NMscanData(NMdata_filepath("examples/nonmem/xgxr001.lst")) : Input data
#> columns will be appended to output data. However, column(s) were identified as
#> unique identifiers, present in both input and output data. If this column or
#> one of these columns is not modified by the Nonmem run, consider using this in
#> col.row for a robust merge of input and output data. Candidate columns: ROW To
#> skip this check, please specify either col.row (recommended) or
#> cbind.by.filters.

#> Model:  xgxr001 
#> 
#> Tables, number of columns in tables, and their detail level:
#>            table used/total level
#>  xgxr001_res.txt      16/16   row
#>            input      22/22   row
#> 
#> Numbers of ID's and rows in data
#>       N From output tables From input data only
#>  N.rows                905                    0
#>   N.ids                150                    0
#> 
#> Distribution of rows on event types
#>  EVID From output tables
#>     0                755
#>     1                150

class(dat)

#> [1] "NMdata"     "data.frame"

library(tibble)
dat <- NMscanData(NMdata_filepath("examples/nonmem/xgxr001.lst"),as.fun=tibble::as_tibble)

#> In NMscanData(NMdata_filepath("examples/nonmem/xgxr001.lst"), as.fun =
#> tibble::as_tibble) : Input data columns will be appended to output data.
#> However, column(s) were identified as unique identifiers, present in both
#> input and output data. If this column or one of these columns is not modified
#> by the Nonmem run, consider using this in col.row for a robust merge of input
#> and output data. Candidate columns: ROW To skip this check, please specify
#> either col.row (recommended) or cbind.by.filters.

#> Model:  xgxr001 
#> 
#> Tables, number of columns in tables, and their detail level:
#>            table used/total level
#>  xgxr001_res.txt      16/16   row
#>            input      22/22   row
#> 
#> Numbers of ID's and rows in data
#>       N From output tables From input data only
#>  N.rows                905                    0
#>   N.ids                150                    0
#> 
#> Distribution of rows on event types
#>  EVID From output tables
#>     0                755
#>     1                150

class(dat)

#> [1] "NMdata"     "tbl_df"     "tbl"        "data.frame"