Read an output table file from NONMEM

Read a table generated by a $TABLE statement in Nonmem. Generally, these files cannot be read by read.table or similar because formatting depends on options in the $TABLE statement, and because Nonmem sometimes includes extra lines in the output that have to be filtered out. NMreadTab can do this automatically based on the table file alone.

NMreadTab(file, tab.count = TRUE, quiet, as.fun, ...)

Arguments

file
path to NONMEM table file

tab.count
Nonmem includes a counter of tables in the written data files. These are often not useful. However, if tab.count is TRUE (default), this will be carried forward and added as a column called TABLENO.

quiet
logical stating whether or not information is prignted about what is being done. Default can be configured using NMdataConf.

as.fun
The default is to return data as a data.frame. Pass a function (say tibble::as_tibble) in as.fun to convert to something else. If data.tables are wanted, use as.fun="data.table". The default can be configured using NMdataConf.

...
Arguments passed to fread.

Value

The Nonmem table data.

Details

The actual reading of data is based on data.table::fread. Generally, the function is fast thanks to data.table.

See also