automatically find Nonmem tables and organize data

automatically find Nonmem tables and organize data

NMscanData(
  file,
  col.row,
  cbind.by.filters,
  use.input = TRUE,
  recover.rows = FALSE,
  add.name = "model",
  name,
  file.mod,
  dir.data,
  quiet = FALSE,
  use.rds = TRUE,
  as.fun = NULL,
  col.id = "ID",
  tab.count = FALSE
)

Arguments

file
A nonmem control stream or output file from nonmem (.mod or .lst)

col.row
A column with a unique value for each row. Such a column is recommended to use if possible. See cbind.by.filters and details as well.

cbind.by.filters
If TRUE, Nonmem data filtering is interpreted from lst file (restrictions apply), and after an imitated selection of rows, data columns will be appended to output data. This method relies on interpretation of Nonmem code, and it will not work in advanced use of IGNORE and ACCEPT statements in $INPUT. Consider using col.row instead, if possible. Default is TRUE if col.row is either missing or NULL. However, explicitly specifying cbind.by.filters is recommended if that is the intended behaviour. If not, NMscanData will search for potential columns to merge by and print information about it. See col.row (recommended method) as well.

use.input
Merge with columns in input data? Using this, you don't have to worry about remembering including all relevant variables in the output tables.

recover.rows
Include rows from input data files that do not exist in output tables? This will be added to the $row dataset only, and $run, $id, and $occ datasets are created before this is taken into account. A column called nmout will be TRUE when the row was found in output tables, and FALSE when not. This is still experimental. More testing is needed.

add.name
If a character string, a column of this name will be included in all tables containing the model name. The default is to store this in a column called "model". See argument "name" as well. Set to NULL if not wanted.

name
The model name to be stored if add.name is not NULL. If name is not supplied, the name will be taken from the control stream file name.

file.mod
The input control stream. Default is to look for \"file\" with extension changed to .mod (PSN style). You can also supply the path to the file, or you can provide a function that translates the output file path to the input file path. See dir.data too.

dir.data
The data directory can only be read from the control stream (.mod) and not from the output file (.lst). So if you only have the output file, use dir.data to tell in which directory to find the data file. If dir.data is provided, the .mod file is not used at all.

quiet
The default is to give some information along the way on what data is found. But consider setting this to TRUE for non-interactive use.

use.rds
If an rds file is found with the exact same name (except for .rds instead of say .csv) as the input data file mentioned in the Nonmem control stream, should this be used instead? The default is yes, and NMwriteData will create this by default too.

as.fun
The default is to return data in data.tables. Pass a function in as.fun to convert to something else. If data.frames are wanted, use as.fun=as.data.frame.

col.id
The name of the subject ID variable, default is "ID".

tab.count
Nonmem includes a counter of tables in the written data files. These are often not useful. Especially for NMscanData output it can be meaningless because multiple tables can be combined so this information is not unique across those source tables. However, if tab.count is TRUE (not default), this will be carried forward and added as a column called TABLENO.

Details

This function makes it very easy to collect the data from a Nonmem run.

A useful feature of this function is that it can automatically combine "input" data (the data read by nonmem in $INPUT or $INFILE) with "output" data (tables written by nonmem in $TABLE). There are two implemented methods for doing so. One (the default but not recommended) relies on interpretation of filter (IGNORE and ACCEPT) statements in $INPUT. This will work in most cases, and checks for consistency with Nonmem results. However, the recommended method is using a unique row identifier in both input data and at least one output data file (not a FIRSTONLY or LASTONLY table). Supply the name of this column using the col.row argument.

See also