replace ($)sections of a nonmem control stream

Just give the section name, the new lines and the file path, and the "$section", and the input to Nonmem will be updated.

NMwriteSection(
  file,
  section,
  newlines,
  list.sections,
  newfile,
  backup = TRUE,
  blank.append = TRUE,
  write = TRUE
)

Arguments

file
File path to the model (control stream) to edit.

section
The name of the section to update. Example: section="EST" to edit the sections starting by $EST. See NMreadSection

newlines
The new text (including "$SECTION"). Better be broken into lines in a character vector since this is simply past to writeLines.

list.sections
Named list of new sections, each element containing a section. Names must be section names, contents of each element are the new section lines for each section.

newfile
path to new run. If missing, path is used. If NULL, output is returned rather than written.

backup
In case you are overwriting the old file, do you want to backup the file (to say, backup_run001.mod)?

blank.append
Append a blank line to output?

write
Default is to write to file. If write=FALSE, NMwriteSection returns the resulting input.txt without writing it. to disk? Default is FALSE.

Details

The new file will be written with unix-style line endings.

See also

Examples

newlines <- "$EST POSTHOC INTERACTION METHOD=1 NOABORT PRINT=5 MAXEVAL=9999 SIG=3"
NMwriteSection(file=system.file("examples/nonmem/xgxr001.mod", package = "NMdata"),
section="EST", newlines=newlines,newfile=NULL)

#>  [1] "$PROBLEM    PK"                                                            
#>  [2] ""                                                                          
#>  [3] ";@ Variables 17/20 @;"                                                     
#>  [4] "$INPUT ROW ID NOMTIME TIME EVID CMT AMT DV FLAG STUDY"                     
#>  [5] "BLQ CYCLE DOSE PART PROFDAY PROFTIME WEIGHTB eff0"                         
#>  [6] ""                                                                          
#>  [7] "$DATA     ../data/xgxr1.csv IGNORE=@ IGNORE=(FLAG.NE.0)"                   
#>  [8] ""                                                                          
#>  [9] "$SUBROUTINE ADVAN4 TRANS4"                                                 
#> [10] ""                                                                          
#> [11] "$PK"                                                                       
#> [12] "TVKA=THETA(1)"                                                             
#> [13] "TVV2=THETA(2)"                                                             
#> [14] "TVCL=THETA(3)"                                                             
#> [15] "TVV3=THETA(4)"                                                             
#> [16] "TVQ=THETA(5)"                                                              
#> [17] "                   "                                                       
#> [18] "KA=TVKA*EXP(ETA(1))"                                                       
#> [19] "V2=TVV2*EXP(ETA(2))"                                                       
#> [20] "CL=TVCL*EXP(ETA(3))"                                                       
#> [21] "V3=TVV3*EXP(ETA(4))"                                                       
#> [22] "Q=TVQ*EXP(ETA(5))"                                                         
#> [23] ""                                                                          
#> [24] "$ERROR"                                                                    
#> [25] "  IPRED=F"                                                                 
#> [26] "  IRES=DV-IPRED"                                                           
#> [27] ""                                                                          
#> [28] "  IF (IPRED.GT.1) THEN"                                                    
#> [29] "    W = SQRT(IPRED**2*SIGMA(1,1)**2 + SIGMA(2,2)**2)"                      
#> [30] "  ELSE"                                                                    
#> [31] "    W=1"                                                                   
#> [32] "  ENDIF"                                                                   
#> [33] ""                                                                          
#> [34] "  IWRES=IRES/W"                                                            
#> [35] "  Y=F+F*ERR(1)+ERR(2)"                                                     
#> [36] ""                                                                          
#> [37] ";-----------------------INITIAL ESTIMATES---------------------------------"
#> [38] "$THETA  (0,0.9)           ; POPKA"                                         
#> [39] "$THETA  (0,4.2)             ; POPCL"                                       
#> [40] "$THETA  (0,3.6)             ; POPV2"                                       
#> [41] "$THETA  (0,5.81)             ; POPV3"                                      
#> [42] "$THETA  (0,3.44)           ; POPQ "                                        
#> [43] ""                                                                          
#> [44] "$OMEGA 0.1"                                                                
#> [45] "$OMEGA 0.1"                                                                
#> [46] "$OMEGA 0.1"                                                                
#> [47] "$OMEGA 0 FIX"                                                              
#> [48] "$OMEGA 0 FIX"                                                              
#> [49] ""                                                                          
#> [50] "$SIGMA .2"                                                                 
#> [51] "$SIGMA .2"                                                                 
#> [52] ""                                                                          
#> [53] "$EST POSTHOC INTERACTION METHOD=1 NOABORT PRINT=5 MAXEVAL=9999 SIG=3"      
#> [54] ""                                                                          
#> [55] "$TABLE ROW   TVKA   TVV2     TVCL  TVV3  TVQ  KA  V2  CL  V3  Q"           
#> [56] "       PRED  IPRED  NOPRINT  FILE=xgxr001_res.txt"