replace ($)sections of a nonmem control stream

Just give the section name, the new lines and the file path, and the "$section", and the input to Nonmem will be updated.

NMwriteSection(
  file,
  section,
  newlines,
  list.sections,
  newfile,
  backup = TRUE,
  blank.append = TRUE,
  test = FALSE
)

Arguments

file
File path to the model (control stream) to edit.

section
The name of the section to update. Example: section="EST" to edit the sections starting by $EST. See NMreadSection

newlines
The new text (including "$SECTION"). Better be broken into lines in a character vector since this is simply past to writeLines.

list.sections
Named list of new sections, each element containing a section. Names must be section names, contents of each element are the new section lines for each section.

newfile
path to new run. If missing, path is used. If NULL, output is returned rather than written.

backup
In case you are overwriting the old file, do you want to backup the file (to say, backup_run001.mod)?

blank.append
Append a blank line to output?

test
Want to see the resulting input.txt and not write it to disk? Default is FALSE.

Details

The new file will be written with unix-style line endings.

See also

Examples

newlines <- "$EST POSTHOC INTERACTION METHOD=1 NOABORT PRINT=5 MAXEVAL=9999 SIG=3"
NMwriteSection(file=system.file("examples/nonmem/run001.mod", package = "NMdata"),
section="EST", newlines=newlines,newfile=NULL)

#>   [1] "$PROBLEM    PK"                                                                                                       
#>   [2] "$INPUT ROW ID NOMTIME TIME EVID CMT AMT DV BLQ CYCLE DOSE FLAG PART PROFDAY PROFTIME STUDY WEIGHTB"                   
#>   [3] ""                                                                                                                     
#>   [4] "$DATA      PK.csv IGNORE=@ IGNORE=(PKFL.EQ.0)"                                                                        
#>   [5] "            IGNORE=(TRTPN.EQ.1) IGNORE=(TRTPN.EQ.3)"                                                                  
#>   [6] "            IGNORE=(TRTPN.EQ.4)"                                                                                      
#>   [7] "$SUBROUTINE ADVAN6 TOL=6"                                                                                             
#>   [8] "$MODEL      NCOMP=4 COMP=(cent) ;; central compartment parent"                                                        
#>   [9] "            COMP=(centmet) ;; central compartment metabolite"                                                         
#>  [10] "            COMP=(peri) ;; peripheral compartment parent"                                                             
#>  [11] "            COMP=(perimet) ;; peripheral compartment metabolite"                                                      
#>  [12] ""                                                                                                                     
#>  [13] ";--------------------------"                                                                                          
#>  [14] "$PK"                                                                                                                  
#>  [15] ""                                                                                                                     
#>  [16] "\"FIRST"                                                                                                              
#>  [17] "\" COMMON/PRCOMG/ IDUM1,IDUM2,IMAX,IDUM4,IDUM5"                                                                       
#>  [18] "\"INTEGER IDUM1,IDUM2,IMAX,IDUM4,IDUM5"                                                                               
#>  [19] "\"IMAX=1000000"                                                                                                       
#>  [20] ""                                                                                                                     
#>  [21] "   ;# Set up dosing times"                                                                                            
#>  [22] "   IF(NEWIND.LE.1) THEN"                                                                                              
#>  [23] "      TDOS= -1000"                                                                                                    
#>  [24] "      PD=0"                                                                                                           
#>  [25] "   ENDIF"                                                                                                             
#>  [26] "   IF(AMT.GT.0) THEN"                                                                                                 
#>  [27] "      TDOS = TIME ; If AMT > 0, set TDOS to TIME"                                                                     
#>  [28] "      PD = AMT    ; If AMT > 0, update PD"                                                                            
#>  [29] "   ENDIF"                                                                                                             
#>  [30] "   TAD2 = TIME-TDOS ; Set time after dose"                                                                            
#>  [31] ""                                                                                                                     
#>  [32] "   E1=EXP(ETA(1))"                                                                                                    
#>  [33] "   E2=EXP(ETA(2))"                                                                                                    
#>  [34] "   E3=EXP(ETA(3))"                                                                                                    
#>  [35] "   E4=EXP(ETA(4))"                                                                                                    
#>  [36] "   E5=EXP(ETA(5))"                                                                                                    
#>  [37] "   E6=EXP(ETA(6))"                                                                                                    
#>  [38] "   E7=EXP(ETA(7))"                                                                                                    
#>  [39] ""                                                                                                                     
#>  [40] ""                                                                                                                     
#>  [41] "   TVCL=THETA(1)"                                                                                                     
#>  [42] "   CL=TVCL*E1"                                                                                                        
#>  [43] ""                                                                                                                     
#>  [44] "   TVV1=THETA(2)"                                                                                                     
#>  [45] "   V1=TVV1"                                                                                                           
#>  [46] ""                                                                                                                     
#>  [47] "   TVV2=THETA(3)"                                                                                                     
#>  [48] "   V2=TVV2"                                                                                                           
#>  [49] ""                                                                                                                     
#>  [50] "   TVQ1=THETA(4)"                                                                                                     
#>  [51] "   Q1=TVQ1*E2"                                                                                                        
#>  [52] ""                                                                                                                     
#>  [53] "   TVCLM=THETA(8)"                                                                                                    
#>  [54] "   CLM=TVCLM*E3"                                                                                                      
#>  [55] ""                                                                                                                     
#>  [56] "   TVVM1=THETA(9)"                                                                                                    
#>  [57] "   VM1=TVVM1"                                                                                                         
#>  [58] ""                                                                                                                     
#>  [59] "   TVVM2=THETA(10)"                                                                                                   
#>  [60] "   VM2=TVVM2*E4"                                                                                                      
#>  [61] ""                                                                                                                     
#>  [62] "   TVQ2=THETA(11)"                                                                                                    
#>  [63] "   Q2=TVQ2"                                                                                                           
#>  [64] ""                                                                                                                     
#>  [65] "   MTT=THETA(6)*E5"                                                                                                   
#>  [66] "   NT=1+THETA(15)*E6"                                                                                                 
#>  [67] "   KTR=NT/MTT"                                                                                                        
#>  [68] "   NN=NT-1"                                                                                                           
#>  [69] ""                                                                                                                     
#>  [70] "   ;;-- Approximation of log(factorial(NN))"                                                                          
#>  [71] "   IF(NN.GE.0.558) THEN"                                                                                              
#>  [72] "      L = LOG(2.5066)+(NN+.5)*LOG(NN)-NN+LOG(1+1/(12*NN))"                                                            
#>  [73] "   ELSE"                                                                                                              
#>  [74] "      L= -0.1171157 + 0.5690*((0.558 - NN)**2) - 0.1122*(0.558 - NN)"                                                 
#>  [75] "   ENDIF"                                                                                                             
#>  [76] ""                                                                                                                     
#>  [77] "   K13=Q1/V1"                                                                                                         
#>  [78] "   K31=Q1/V2"                                                                                                         
#>  [79] ""                                                                                                                     
#>  [80] "   TVfm=0.22"                                                                                                         
#>  [81] "   fm=TVfm"                                                                                                           
#>  [82] ""                                                                                                                     
#>  [83] "   K10=CL*(1-fm)/V1"                                                                                                  
#>  [84] "   K12=fm*CL/V1"                                                                                                      
#>  [85] ""                                                                                                                     
#>  [86] "   K24=Q2/VM1"                                                                                                        
#>  [87] "   K42=Q2/VM2"                                                                                                        
#>  [88] "   K20=CLM/VM1"                                                                                                       
#>  [89] ""                                                                                                                     
#>  [90] "   S1=V1"                                                                                                             
#>  [91] "   S2=VM1"                                                                                                            
#>  [92] "   F1=0"                                                                                                              
#>  [93] "   BIO=THETA(7)*E7"                                                                                                   
#>  [94] ""                                                                                                                     
#>  [95] "   X=0.0000001"                                                                                                       
#>  [96] "$DES"                                                                                                                 
#>  [97] "   IF(T.GE.TDOS) THEN"                                                                                                
#>  [98] "      DADT(1)= BIO*PD*KTR*EXP(NN*LOG(KTR*(T-TDOS) + X) - KTR*(T - TDOS)-L)  - k10*A(1) - k12*A(1) - k13*A(1)+k31*A(3)"
#>  [99] "   ELSE"                                                                                                              
#> [100] "      DADT(1)=  - k12*A(1) - k13*A(1)+k31*A(3)"                                                                       
#> [101] "   ENDIF"                                                                                                             
#> [102] "   DADT(2)=  k12*A(1) - k20*A(2) - k24*A(2)+k42*A(4)"                                                                 
#> [103] "   DADT(3)=  k13*A(1)-k31*A(3)"                                                                                       
#> [104] "   DADT(4)=  k24*A(2)-k42*A(4)"                                                                                       
#> [105] ""                                                                                                                     
#> [106] "$ERROR"                                                                                                               
#> [107] ""                                                                                                                     
#> [108] "   STRT=FLAG"                                                                                                         
#> [109] ""                                                                                                                     
#> [110] "   ;;Separate residual error for parent and the active metabolite"                                                    
#> [111] ""                                                                                                                     
#> [112] "   IF(FLAG.LE.1) THEN"                                                                                                
#> [113] "      IPRED=A(1)/S1"                                                                                                  
#> [114] "      W=sqrt( (theta(5)*IPRED)**2 + theta(13)**2 )  ;;prop and add  error parent"                                     
#> [115] "      Y=IPRED+W*EPS(1)"                                                                                               
#> [116] "   ELSE"                                                                                                              
#> [117] "      IPRED=A(2)/S2"                                                                                                  
#> [118] "      W=sqrt( (theta(12)*IPRED)**2 + theta(14)**2 )  ;;prop and add  error metabolite"                                
#> [119] "      Y=IPRED+W*EPS(1)"                                                                                               
#> [120] "   ENDIF"                                                                                                             
#> [121] ""                                                                                                                     
#> [122] "   CP=A(1)/S1"                                                                                                        
#> [123] "   CM=A(2)/S2"                                                                                                        
#> [124] ""                                                                                                                     
#> [125] "   DEL=0"                                                                                                             
#> [126] "   IF (W.EQ.0) DEL=1"                                                                                                 
#> [127] "   IRES=DV-IPRED"                                                                                                     
#> [128] "   IWRES=IRES/(W+DEL)"                                                                                                
#> [129] ""                                                                                                                     
#> [130] ";--------------------------"                                                                                          
#> [131] "$THETA  (0,16.5964) ; 1 CL L/h parent"                                                                                
#> [132] " (0,144.673) ; 2 V1 L parent"                                                                                         
#> [133] " (0,68.9777) ; 3 V2 L parent"                                                                                         
#> [134] " (0,15.7851) ; 4 Q1 parent"                                                                                           
#> [135] " (0,0.36446) ; 5  ERROR parent"                                                                                       
#> [136] " (0,1.00) ; 6 MTT"                                                                                                    
#> [137] " 1 FIX ; 7 F1"                                                                                                        
#> [138] " (0,10.387) ; 8 CLM metabolite"                                                                                       
#> [139] " (0,7.53503) ; 9 VM1 L metabolite"                                                                                    
#> [140] " (0,43.3393) ; 10 VM2 L metabolite"                                                                                   
#> [141] " (0,4.55635) ; 11 Q2 metabolite"                                                                                      
#> [142] " (0,0.279073) ; 12 ERROR metabolite"                                                                                  
#> [143] " (0,3) ; 13 Add ERROR parent"                                                                                         
#> [144] " (0,3) ; 14 Add ERROR metabolite"                                                                                     
#> [145] " (1,5,12) ; 15 NTR"                                                                                                   
#> [146] ";--------------------------"                                                                                          
#> [147] "$OMEGA  0.0558419  ; 1 CL parent"                                                                                     
#> [148] "$OMEGA  0.193658  ; 2 Q1 parent"                                                                                      
#> [149] "$OMEGA  0.0857882  ; 3 CL metabolite"                                                                                 
#> [150] "$OMEGA  0  FIX  ; 4 V2 metabolite"                                                                                    
#> [151] "$OMEGA  0.321524  ;      5 KTR"                                                                                       
#> [152] "$OMEGA  0  FIX  ;      6 NTR"                                                                                         
#> [153] "$OMEGA  0.0975457  ; 7 Bioavailability"                                                                               
#> [154] ";--------------------------"                                                                                          
#> [155] "$SIGMA  1  FIX"                                                                                                       
#> [156] ";--------------------------"                                                                                          
#> [157] "$EST POSTHOC INTERACTION METHOD=1 NOABORT PRINT=5 MAXEVAL=9999 SIG=3"                                                 
#> [158] ""                                                                                                                     
#> [159] "$COVARIANCE PRINT=E"                                                                                                  
#> [160] "$TABLE      PKFL DV ID FLAG PERIOD TAD TIME AMT EVID CMT MDV TRTPN"                                                   
#> [161] "            IPRED IWRES RES CWRES CL V1 V2 Q1 F1 CLM fm VM1 VM2 Q2 KTR"                                               
#> [162] "            MTT NT ETA1 ETA2 ETA3 ETA4 ETA5 ETA6 ETA7 NOPRINT"                                                        
#> [163] "            ONEHEADER file=run001.txt"