Extract columns that do not vary within variables in data

Extract columns that do not vary within variables in data

findCovs(data, cols.id = NULL, as.fun = NULL)

Arguments

data
data.frame in which to look for covariates

cols.id
covariates will be searched for in combinations of values in these columns. Often cols.id will be either empty or ID. But it can also be both say c("ID","DRUG") or c("ID","TRT").

as.fun
The default is to return a data.table if data is a data.table and return a data.frame in all other cases. Pass a function in as.fun to convert to something else. See ?runAsFun.

Value

a data set with one observation per combination of values of variables listed in cols.id.

See also

Examples

dat <- NMscanData(NMdata_filepath("examples/nonmem/xgxr001.lst"))

#> In NMscanData(NMdata_filepath("examples/nonmem/xgxr001.lst")) : Input data
#> columns will be appended to output data. However, column(s) were identified as
#> unique identifiers, present in both input and output data. If this column or
#> one of these columns is not modified by the Nonmem run, consider using this in
#> col.row for a robust merge of input and output data. Candidate columns: ROW To
#> skip this check, please specify either col.row (recommended) or
#> cbind.by.filters.

#> Model:  xgxr001 
#> 
#> Tables, number of columns in tables, and their detail level:
#>            table used/total level
#>  xgxr001_res.txt      16/16   row
#>            input      22/22   row
#> 
#> Numbers of ID's and rows in data
#>       N From output tables
#>  N.rows                905
#>   N.ids                150
#> 
#> Distribution of rows on event types
#>  EVID From output tables
#>     0                755
#>     1                150

### very common use
findCovs(dat,cols.id="ID")

#>      ID    TVKA  TVV2    TVCL    TVV3     TVQ       KA       V2      CL      V3
#> 1    31 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 2    32 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 3    33 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 4    34 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 5    35 0.18116 0.042 0.72491 0.17857 8813700 0.200980 0.044867 0.72487 0.17857
#> 6    36 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 7    37 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 8    38 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 9    39 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 10   40 0.18116 0.042 0.72491 0.17857 8813700 0.198940 0.044708 0.72487 0.17857
#> 11   41 0.18116 0.042 0.72491 0.17857 8813700 0.197620 0.044186 0.72488 0.17857
#> 12   42 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 13   43 0.18116 0.042 0.72491 0.17857 8813700 0.210460 0.046150 0.72486 0.17857
#> 14   44 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 15   45 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 16   46 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 17   47 0.18116 0.042 0.72491 0.17857 8813700 0.193710 0.044405 0.72487 0.17857
#> 18   48 0.18116 0.042 0.72491 0.17857 8813700 0.193420 0.044558 0.72487 0.17857
#> 19   49 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 20   50 0.18116 0.042 0.72491 0.17857 8813700 0.199420 0.044395 0.72488 0.17857
#> 21   51 0.18116 0.042 0.72491 0.17857 8813700 0.232920 0.051853 0.72478 0.17857
#> 22   52 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 23   53 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 24   54 0.18116 0.042 0.72491 0.17857 8813700 0.197710 0.045296 0.72486 0.17857
#> 25   55 0.18116 0.042 0.72491 0.17857 8813700 0.196600 0.045040 0.72486 0.17857
#> 26   56 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 27   57 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 28   58 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 29   59 0.18116 0.042 0.72491 0.17857 8813700 0.236010 0.052070 0.72478 0.17857
#> 30   60 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000 0.72491 0.17857
#> 31   61 0.18116 0.042 0.72491 0.17857 8813700 0.226490 0.050277 0.72479 0.17857
#> 32   62 0.18116 0.042 0.72491 0.17857 8813700 0.182800 0.042699 0.72490 0.17857
#> 33   63 0.18116 0.042 0.72491 0.17857 8813700 0.157820 0.038191 0.72496 0.17857
#> 34   64 0.18116 0.042 0.72491 0.17857 8813700 0.167120 0.040147 0.72493 0.17857
#> 35   65 0.18116 0.042 0.72491 0.17857 8813700 0.175540 0.041752 0.72490 0.17857
#> 36   66 0.18116 0.042 0.72491 0.17857 8813700 0.156170 0.038461 0.72495 0.17857
#> 37   67 0.18116 0.042 0.72491 0.17857 8813700 0.154820 0.037750 0.72497 0.17857
#> 38   68 0.18116 0.042 0.72491 0.17857 8813700 0.207850 0.045044 0.72487 0.17857
#> 39   69 0.18116 0.042 0.72491 0.17857 8813700 0.200940 0.043803 0.72492 0.17857
#> 40   70 0.18116 0.042 0.72491 0.17857 8813700 0.177070 0.041886 0.72490 0.17857
#> 41   71 0.18116 0.042 0.72491 0.17857 8813700 0.197000 0.044794 0.72488 0.17857
#> 42   72 0.18116 0.042 0.72491 0.17857 8813700 0.193850 0.049011 0.72481 0.17857
#> 43   73 0.18116 0.042 0.72491 0.17857 8813700 0.192760 0.049967 0.72477 0.17857
#> 44   74 0.18116 0.042 0.72491 0.17857 8813700 0.238040 0.053145 0.72475 0.17857
#> 45   75 0.18116 0.042 0.72491 0.17857 8813700 0.224240 0.051677 0.72475 0.17857
#> 46   76 0.18116 0.042 0.72491 0.17857 8813700 0.184790 0.042526 0.72490 0.17857
#> 47   77 0.18116 0.042 0.72491 0.17857 8813700 0.243810 0.059397 0.72466 0.17857
#> 48   78 0.18116 0.042 0.72491 0.17857 8813700 0.154970 0.038192 0.72495 0.17857
#> 49   79 0.18116 0.042 0.72491 0.17857 8813700 0.183190 0.040635 0.72495 0.17857
#> 50   80 0.18116 0.042 0.72491 0.17857 8813700 0.168150 0.040651 0.72491 0.17857
#> 51   81 0.18116 0.042 0.72491 0.17857 8813700 0.306360 0.059155 0.72476 0.17857
#> 52   82 0.18116 0.042 0.72491 0.17857 8813700 0.164380 0.039374 0.72493 0.17857
#> 53   83 0.18116 0.042 0.72491 0.17857 8813700 0.217370 0.049527 0.72480 0.17857
#> 54   84 0.18116 0.042 0.72491 0.17857 8813700 0.195560 0.042991 0.72490 0.17857
#> 55   85 0.18116 0.042 0.72491 0.17857 8813700 0.242330 0.053778 0.72476 0.17857
#> 56   86 0.18116 0.042 0.72491 0.17857 8813700 0.198640 0.046318 0.72482 0.17857
#> 57   87 0.18116 0.042 0.72491 0.17857 8813700 0.202980 0.048004 0.72481 0.17857
#> 58   88 0.18116 0.042 0.72491 0.17857 8813700 0.196660 0.046519 0.72484 0.17857
#> 59   89 0.18116 0.042 0.72491 0.17857 8813700 0.180460 0.041720 0.72491 0.17857
#> 60   90 0.18116 0.042 0.72491 0.17857 8813700 0.176070 0.041943 0.72489 0.17857
#> 61   91 0.18116 0.042 0.72491 0.17857 8813700 0.304040 0.058857 0.72466 0.17857
#> 62   92 0.18116 0.042 0.72491 0.17857 8813700 0.135840 0.041109 0.72484 0.17857
#> 63   93 0.18116 0.042 0.72491 0.17857 8813700 0.196360 0.032508 0.72516 0.17857
#> 64   94 0.18116 0.042 0.72491 0.17857 8813700 0.152420 0.035164 0.72497 0.17857
#> 65   95 0.18116 0.042 0.72491 0.17857 8813700 0.160710 0.040626 0.72485 0.17857
#> 66   96 0.18116 0.042 0.72491 0.17857 8813700 0.159820 0.038148 0.72487 0.17857
#> 67   97 0.18116 0.042 0.72491 0.17857 8813700 0.164980 0.033916 0.72504 0.17857
#> 68   98 0.18116 0.042 0.72491 0.17857 8813700 0.111950 0.063151 0.72463 0.17857
#> 69   99 0.18116 0.042 0.72491 0.17857 8813700 0.191610 0.045673 0.72480 0.17857
#> 70  100 0.18116 0.042 0.72491 0.17857 8813700 0.139770 0.032785 0.72504 0.17857
#> 71  101 0.18116 0.042 0.72491 0.17857 8813700 0.218490 0.038332 0.72499 0.17857
#> 72  102 0.18116 0.042 0.72491 0.17857 8813700 0.284460 0.044049 0.72490 0.17857
#> 73  103 0.18116 0.042 0.72491 0.17857 8813700 0.111030 0.042973 0.72490 0.17857
#> 74  104 0.18116 0.042 0.72491 0.17857 8813700 0.146970 0.037805 0.72495 0.17857
#> 75  105 0.18116 0.042 0.72491 0.17857 8813700 0.311480 0.048335 0.72488 0.17857
#> 76  106 0.18116 0.042 0.72491 0.17857 8813700 0.140450 0.035079 0.72502 0.17857
#> 77  107 0.18116 0.042 0.72491 0.17857 8813700 0.192910 0.035892 0.72504 0.17857
#> 78  108 0.18116 0.042 0.72491 0.17857 8813700 0.150810 0.036938 0.72502 0.17857
#> 79  109 0.18116 0.042 0.72491 0.17857 8813700 0.122350 0.045355 0.72485 0.17857
#> 80  110 0.18116 0.042 0.72491 0.17857 8813700 0.205140 0.036559 0.72508 0.17857
#> 81  111 0.18116 0.042 0.72491 0.17857 8813700 0.298110 0.041774 0.72494 0.17857
#> 82  112 0.18116 0.042 0.72491 0.17857 8813700 0.235970 0.043731 0.72487 0.17857
#> 83  113 0.18116 0.042 0.72491 0.17857 8813700 0.200640 0.050461 0.72475 0.17857
#> 84  114 0.18116 0.042 0.72491 0.17857 8813700 0.163250 0.044300 0.72477 0.17857
#> 85  115 0.18116 0.042 0.72491 0.17857 8813700 0.191250 0.035263 0.72500 0.17857
#> 86  116 0.18116 0.042 0.72491 0.17857 8813700 0.141670 0.033440 0.72501 0.17857
#> 87  117 0.18116 0.042 0.72491 0.17857 8813700 0.107440 0.030178 0.72509 0.17857
#> 88  118 0.18116 0.042 0.72491 0.17857 8813700 0.147590 0.031952 0.72510 0.17857
#> 89  119 0.18116 0.042 0.72491 0.17857 8813700 0.196160 0.034625 0.72503 0.17857
#> 90  120 0.18116 0.042 0.72491 0.17857 8813700 0.117460 0.032440 0.72506 0.17857
#> 91  121 0.18116 0.042 0.72491 0.17857 8813700 0.220590 0.038723 0.72486 0.17857
#> 92  122 0.18116 0.042 0.72491 0.17857 8813700 0.136500 0.038012 0.72480 0.17857
#> 93  123 0.18116 0.042 0.72491 0.17857 8813700 0.160980 0.042487 0.72488 0.17857
#> 94  124 0.18116 0.042 0.72491 0.17857 8813700 0.219600 0.033770 0.72500 0.17857
#> 95  125 0.18116 0.042 0.72491 0.17857 8813700 0.153720 0.052751 0.72482 0.17857
#> 96  126 0.18116 0.042 0.72491 0.17857 8813700 0.295100 0.042108 0.72494 0.17857
#> 97  127 0.18116 0.042 0.72491 0.17857 8813700 0.075578 0.086706 0.72434 0.17857
#> 98  128 0.18116 0.042 0.72491 0.17857 8813700 0.139970 0.039734 0.72512 0.17857
#> 99  129 0.18116 0.042 0.72491 0.17857 8813700 0.204490 0.044726 0.72482 0.17857
#> 100 130 0.18116 0.042 0.72491 0.17857 8813700 0.190000 0.042021 0.72483 0.17857
#> 101 131 0.18116 0.042 0.72491 0.17857 8813700 0.084683 0.065548 0.72459 0.17857
#> 102 132 0.18116 0.042 0.72491 0.17857 8813700 0.103500 0.069220 0.72460 0.17857
#> 103 133 0.18116 0.042 0.72491 0.17857 8813700 0.205740 0.032194 0.72504 0.17857
#> 104 134 0.18116 0.042 0.72491 0.17857 8813700 0.614680 0.036342 0.72507 0.17857
#> 105 135 0.18116 0.042 0.72491 0.17857 8813700 0.302330 0.041101 0.72483 0.17857
#> 106 136 0.18116 0.042 0.72491 0.17857 8813700 0.090507 0.041428 0.72479 0.17857
#> 107 137 0.18116 0.042 0.72491 0.17857 8813700 0.331930 0.031867 0.72518 0.17857
#> 108 138 0.18116 0.042 0.72491 0.17857 8813700 0.119460 0.037087 0.72494 0.17857
#> 109 139 0.18116 0.042 0.72491 0.17857 8813700 0.091146 0.075929 0.72449 0.17857
#> 110 140 0.18116 0.042 0.72491 0.17857 8813700 0.308080 0.040089 0.72481 0.17857
#> 111 141 0.18116 0.042 0.72491 0.17857 8813700 0.087996 0.047212 0.72467 0.17857
#> 112 142 0.18116 0.042 0.72491 0.17857 8813700 0.214520 0.028516 0.72523 0.17857
#> 113 143 0.18116 0.042 0.72491 0.17857 8813700 0.110090 0.063183 0.72466 0.17857
#> 114 144 0.18116 0.042 0.72491 0.17857 8813700 0.142500 0.058911 0.72470 0.17857
#> 115 145 0.18116 0.042 0.72491 0.17857 8813700 0.221820 0.046501 0.72497 0.17857
#> 116 146 0.18116 0.042 0.72491 0.17857 8813700 0.181220 0.051077 0.72488 0.17857
#> 117 147 0.18116 0.042 0.72491 0.17857 8813700 0.167850 0.036629 0.72486 0.17857
#> 118 148 0.18116 0.042 0.72491 0.17857 8813700 0.144050 0.048443 0.72480 0.17857
#> 119 149 0.18116 0.042 0.72491 0.17857 8813700 0.330020 0.031505 0.72511 0.17857
#> 120 150 0.18116 0.042 0.72491 0.17857 8813700 0.240580 0.026797 0.72508 0.17857
#> 121 151 0.18116 0.042 0.72491 0.17857 8813700 0.074433 0.092395 0.72424 0.17857
#> 122 152 0.18116 0.042 0.72491 0.17857 8813700 0.116810 0.051465 0.72468 0.17857
#> 123 153 0.18116 0.042 0.72491 0.17857 8813700 0.422940 0.034489 0.72501 0.17857
#> 124 154 0.18116 0.042 0.72491 0.17857 8813700 0.096618 0.051255 0.72494 0.17857
#> 125 155 0.18116 0.042 0.72491 0.17857 8813700 0.151840 0.035464 0.72498 0.17857
#> 126 156 0.18116 0.042 0.72491 0.17857 8813700 0.370540 0.022131 0.72508 0.17857
#> 127 157 0.18116 0.042 0.72491 0.17857 8813700 0.170190 0.066922 0.72488 0.17857
#> 128 158 0.18116 0.042 0.72491 0.17857 8813700 0.125960 0.056415 0.72457 0.17857
#> 129 159 0.18116 0.042 0.72491 0.17857 8813700 0.133700 0.047473 0.72475 0.17857
#> 130 160 0.18116 0.042 0.72491 0.17857 8813700 0.406330 0.020624 0.72521 0.17857
#> 131 161 0.18116 0.042 0.72491 0.17857 8813700 0.144910 0.069496 0.72523 0.17857
#> 132 162 0.18116 0.042 0.72491 0.17857 8813700 0.250630 0.020414 0.72526 0.17857
#> 133 163 0.18116 0.042 0.72491 0.17857 8813700 0.192570 0.024115 0.72518 0.17857
#> 134 164 0.18116 0.042 0.72491 0.17857 8813700 0.123500 0.061082 0.72469 0.17857
#> 135 165 0.18116 0.042 0.72491 0.17857 8813700 0.401870 0.020528 0.72531 0.17857
#> 136 166 0.18116 0.042 0.72491 0.17857 8813700 0.174450 0.038948 0.72496 0.17857
#> 137 167 0.18116 0.042 0.72491 0.17857 8813700 0.282040 0.033692 0.72506 0.17857
#> 138 168 0.18116 0.042 0.72491 0.17857 8813700 0.285040 0.040023 0.72495 0.17857
#> 139 169 0.18116 0.042 0.72491 0.17857 8813700 0.116410 0.051393 0.72463 0.17857
#> 140 170 0.18116 0.042 0.72491 0.17857 8813700 0.297140 0.021851 0.72537 0.17857
#> 141 171 0.18116 0.042 0.72491 0.17857 8813700 0.091617 0.054356 0.72470 0.17857
#> 142 172 0.18116 0.042 0.72491 0.17857 8813700 0.355320 0.022144 0.72537 0.17857
#> 143 173 0.18116 0.042 0.72491 0.17857 8813700 0.324240 0.038398 0.72517 0.17857
#> 144 174 0.18116 0.042 0.72491 0.17857 8813700 0.135420 0.057353 0.72465 0.17857
#> 145 175 0.18116 0.042 0.72491 0.17857 8813700 0.312270 0.036866 0.72518 0.17857
#> 146 176 0.18116 0.042 0.72491 0.17857 8813700 0.112360 0.058950 0.72475 0.17857
#> 147 177 0.18116 0.042 0.72491 0.17857 8813700 0.134660 0.041848 0.72495 0.17857
#> 148 178 0.18116 0.042 0.72491 0.17857 8813700 0.158520 0.051512 0.72486 0.17857
#> 149 179 0.18116 0.042 0.72491 0.17857 8813700 0.148380 0.067208 0.72512 0.17857
#> 150 180 0.18116 0.042 0.72491 0.17857 8813700 0.221310 0.022291 0.72520 0.17857
#>           Q nmout BLQ CYCLE DOSE FLAG PART PROFDAY STUDY WEIGHTB   eff0
#> 1   8813700  TRUE   0     1    3    0    1       1     1  87.031 56.461
#> 2   8813700  TRUE   0     1    3    0    1       1     1 100.620 45.096
#> 3   8813700  TRUE   0     1    3    0    1       1     1 112.990 49.189
#> 4   8813700  TRUE   0     1    3    0    1       1     1 144.130 56.239
#> 5   8813700  TRUE   0     1    3    0    1       1     1  80.384 58.419
#> 6   8813700  TRUE   0     1    3    0    1       1     1 120.860 58.163
#> 7   8813700  TRUE   0     1    3    0    1       1     1 138.440 42.122
#> 8   8813700  TRUE   0     1    3    0    1       1     1 123.350 59.180
#> 9   8813700  TRUE   0     1    3    0    1       1     1 137.080 47.222
#> 10  8813700  TRUE   0     1    3    0    1       1     1 134.090 49.498
#> 11  8813700  TRUE   0     1    3    0    1       1     1  80.018 41.362
#> 12  8813700  TRUE   0     1    3    0    1       1     1 142.900 54.624
#> 13  8813700  TRUE   0     1    3    0    1       1     1  97.637 51.154
#> 14  8813700  TRUE   0     1    3    0    1       1     1 105.310 54.303
#> 15  8813700  TRUE   0     1    3    0    1       1     1  83.345 59.970
#> 16  8813700  TRUE   0     1    3    0    1       1     1 148.540 40.533
#> 17  8813700  TRUE   0     1    3    0    1       1     1 121.060 42.542
#> 18  8813700  TRUE   0     1    3    0    1       1     1  97.761 51.290
#> 19  8813700  TRUE   0     1    3    0    1       1     1 144.240 45.758
#> 20  8813700  TRUE   0     1    3    0    1       1     1 111.400 44.106
#> 21  8813700  TRUE   0     1    3    0    1       1     1 144.870 42.832
#> 22  8813700  TRUE   0     1    3    0    1       1     1 136.980 48.142
#> 23  8813700  TRUE   0     1    3    0    1       1     1 124.950 52.037
#> 24  8813700  TRUE   0     1    3    0    1       1     1 139.260 54.637
#> 25  8813700  TRUE   0     1    3    0    1       1     1 127.350 57.822
#> 26  8813700  TRUE   0     1    3    0    1       1     1 136.260 48.192
#> 27  8813700  TRUE   0     1    3    0    1       1     1 139.740 45.235
#> 28  8813700  TRUE   0     1    3    0    1       1     1  90.185 51.918
#> 29  8813700  TRUE   0     1    3    0    1       1     1 142.720 54.874
#> 30  8813700  TRUE   0     1    3    0    1       1     1  91.702 57.376
#> 31  8813700  TRUE   0     1   10    0    1       1     1 133.040 58.744
#> 32  8813700  TRUE   0     1   10    0    1       1     1 104.290 40.081
#> 33  8813700  TRUE   0     1   10    0    1       1     1 107.650 49.956
#> 34  8813700  TRUE   0     1   10    0    1       1     1 141.420 56.587
#> 35  8813700  TRUE   0     1   10    0    1       1     1 113.970 50.127
#> 36  8813700  TRUE   0     1   10    0    1       1     1 112.190 49.280
#> 37  8813700  TRUE   0     1   10    0    1       1     1  87.380 48.628
#> 38  8813700  TRUE   0     1   10    0    1       1     1 126.680 47.701
#> 39  8813700  TRUE   0     1   10    0    1       1     1 110.810 44.421
#> 40  8813700  TRUE   0     1   10    0    1       1     1 142.900 50.431
#> 41  8813700  TRUE   0     1   10    0    1       1     1  88.158 58.193
#> 42  8813700  TRUE   0     1   10    0    1       1     1 115.390 59.961
#> 43  8813700  TRUE   0     1   10    0    1       1     1  91.806 41.836
#> 44  8813700  TRUE   0     1   10    0    1       1     1 111.080 42.848
#> 45  8813700  TRUE   0     1   10    0    1       1     1 120.870 56.245
#> 46  8813700  TRUE   0     1   10    0    1       1     1 102.040 47.031
#> 47  8813700  TRUE   0     1   10    0    1       1     1 143.750 52.876
#> 48  8813700  TRUE   0     1   10    0    1       1     1 148.760 46.962
#> 49  8813700  TRUE   0     1   10    0    1       1     1 128.350 43.855
#> 50  8813700  TRUE   0     1   10    0    1       1     1 143.310 57.924
#> 51  8813700  TRUE   0     1   10    0    1       1     1 145.680 48.497
#> 52  8813700  TRUE   0     1   10    0    1       1     1 125.550 48.890
#> 53  8813700  TRUE   0     1   10    0    1       1     1 124.870 41.692
#> 54  8813700  TRUE   0     1   10    0    1       1     1 127.300 47.241
#> 55  8813700  TRUE   0     1   10    0    1       1     1 139.600 55.084
#> 56  8813700  TRUE   0     1   10    0    1       1     1 147.170 42.582
#> 57  8813700  TRUE   0     1   10    0    1       1     1 112.100 59.316
#> 58  8813700  TRUE   0     1   10    0    1       1     1 122.970 49.694
#> 59  8813700  TRUE   0     1   10    0    1       1     1 101.600 40.517
#> 60  8813700  TRUE   0     1   10    0    1       1     1 126.330 46.436
#> 61  8813700  TRUE   0     1   30    0    1       1     1  91.716 58.817
#> 62  8813700  TRUE   0     1   30    0    1       1     1 101.980 41.972
#> 63  8813700  TRUE   0     1   30    0    1       1     1 131.860 57.696
#> 64  8813700  TRUE   0     1   30    0    1       1     1  84.754 42.902
#> 65  8813700  TRUE   0     1   30    0    1       1     1  83.878 41.517
#> 66  8813700  TRUE   0     1   30    0    1       1     1 121.110 57.137
#> 67  8813700  TRUE   0     1   30    0    1       1     1 133.690 55.847
#> 68  8813700  TRUE   0     1   30    0    1       1     1  80.206 55.302
#> 69  8813700  TRUE   0     1   30    0    1       1     1  96.745 49.018
#> 70  8813700  TRUE   0     1   30    0    1       1     1 109.040 46.163
#> 71  8813700  TRUE   0     1   30    0    1       1     1 142.040 53.020
#> 72  8813700  TRUE   0     1   30    0    1       1     1 111.700 53.996
#> 73  8813700  TRUE   0     1   30    0    1       1     1  92.732 50.635
#> 74  8813700  TRUE   0     1   30    0    1       1     1  83.444 58.264
#> 75  8813700  TRUE   0     1   30    0    1       1     1 121.690 49.004
#> 76  8813700  TRUE   0     1   30    0    1       1     1 133.300 53.440
#> 77  8813700  TRUE   0     1   30    0    1       1     1  84.755 49.630
#> 78  8813700  TRUE   0     1   30    0    1       1     1  95.846 46.434
#> 79  8813700  TRUE   0     1   30    0    1       1     1 113.810 57.663
#> 80  8813700  TRUE   0     1   30    0    1       1     1 116.780 58.513
#> 81  8813700  TRUE   0     1   30    0    1       1     1 138.870 58.687
#> 82  8813700  TRUE   0     1   30    0    1       1     1 114.000 50.922
#> 83  8813700  TRUE   0     1   30    0    1       1     1 114.260 49.663
#> 84  8813700  TRUE   0     1   30    0    1       1     1 105.090 49.476
#> 85  8813700  TRUE   0     1   30    0    1       1     1 129.360 50.889
#> 86  8813700  TRUE   0     1   30    0    1       1     1 109.160 42.200
#> 87  8813700  TRUE   0     1   30    0    1       1     1 116.960 45.204
#> 88  8813700  TRUE   0     1   30    0    1       1     1 119.330 40.657
#> 89  8813700  TRUE   0     1   30    0    1       1     1 130.040 57.173
#> 90  8813700  TRUE   0     1   30    0    1       1     1  95.384 47.696
#> 91  8813700  TRUE   0     1  100    0    1       1     1 136.770 47.184
#> 92  8813700  TRUE   0     1  100    0    1       1     1  94.357 49.776
#> 93  8813700  TRUE   0     1  100    0    1       1     1  81.491 57.013
#> 94  8813700  TRUE   0     1  100    0    1       1     1 115.370 51.224
#> 95  8813700  TRUE   0     1  100    0    1       1     1 148.110 41.539
#> 96  8813700  TRUE   0     1  100    0    1       1     1 148.230 48.961
#> 97  8813700  TRUE   0     1  100    0    1       1     1 133.160 42.732
#> 98  8813700  TRUE   0     1  100    0    1       1     1  82.140 49.704
#> 99  8813700  TRUE   0     1  100    0    1       1     1  88.087 46.169
#> 100 8813700  TRUE   0     1  100    0    1       1     1 142.170 49.395
#> 101 8813700  TRUE   0     1  100    0    1       1     1 117.810 57.872
#> 102 8813700  TRUE   0     1  100    0    1       1     1 132.710 42.860
#> 103 8813700  TRUE   0     1  100    0    1       1     1 114.950 40.940
#> 104 8813700  TRUE   0     1  100    0    1       1     1 145.100 54.023
#> 105 8813700  TRUE   0     1  100    0    1       1     1 117.550 55.676
#> 106 8813700  TRUE   0     1  100    0    1       1     1 117.820 45.134
#> 107 8813700  TRUE   0     1  100    0    1       1     1 133.540 45.305
#> 108 8813700  TRUE   0     1  100    0    1       1     1 119.450 49.917
#> 109 8813700  TRUE   0     1  100    0    1       1     1 149.550 43.930
#> 110 8813700  TRUE   0     1  100    0    1       1     1 112.980 48.596
#> 111 8813700  TRUE   0     1  100    0    1       1     1 130.200 44.755
#> 112 8813700  TRUE   0     1  100    0    1       1     1 147.080 50.895
#> 113 8813700  TRUE   0     1  100    0    1       1     1  96.265 43.053
#> 114 8813700  TRUE   0     1  100    0    1       1     1 132.650 51.302
#> 115 8813700  TRUE   0     1  100    0    1       1     1 118.270 50.193
#> 116 8813700  TRUE   0     1  100    0    1       1     1 114.600 56.089
#> 117 8813700  TRUE   0     1  100    0    1       1     1 141.930 54.312
#> 118 8813700  TRUE   0     1  100    0    1       1     1  99.604 44.034
#> 119 8813700  TRUE   0     1  100    0    1       1     1 111.110 55.695
#> 120 8813700  TRUE   0     1  100    0    1       1     1 111.820 55.035
#> 121 8813700  TRUE   0     1  300    0    1       1     1  88.319 53.035
#> 122 8813700  TRUE   0     1  300    0    1       1     1 121.320 47.391
#> 123 8813700  TRUE   0     1  300    0    1       1     1  93.475 56.259
#> 124 8813700  TRUE   0     1  300    0    1       1     1 147.680 52.824
#> 125 8813700  TRUE   0     1  300    0    1       1     1  96.953 58.300
#> 126 8813700  TRUE   0     1  300    0    1       1     1 140.810 46.160
#> 127 8813700  TRUE   0     1  300    0    1       1     1  94.526 53.894
#> 128 8813700  TRUE   0     1  300    0    1       1     1 146.820 54.044
#> 129 8813700  TRUE   0     1  300    0    1       1     1 145.530 55.699
#> 130 8813700  TRUE   0     1  300    0    1       1     1 127.400 50.455
#> 131 8813700  TRUE   0     1  300    0    1       1     1 101.580 55.400
#> 132 8813700  TRUE   0     1  300    0    1       1     1 102.620 59.111
#> 133 8813700  TRUE   0     1  300    0    1       1     1  85.221 41.015
#> 134 8813700  TRUE   0     1  300    0    1       1     1  84.974 54.064
#> 135 8813700  TRUE   0     1  300    0    1       1     1 124.230 53.337
#> 136 8813700  TRUE   0     1  300    0    1       1     1 101.180 59.792
#> 137 8813700  TRUE   0     1  300    0    1       1     1 119.050 45.097
#> 138 8813700  TRUE   0     1  300    0    1       1     1  95.155 46.504
#> 139 8813700  TRUE   0     1  300    0    1       1     1 124.980 54.155
#> 140 8813700  TRUE   0     1  300    0    1       1     1 102.150 51.862
#> 141 8813700  TRUE   0     1  300    0    1       1     1 136.230 47.295
#> 142 8813700  TRUE   0     1  300    0    1       1     1  80.888 49.242
#> 143 8813700  TRUE   0     1  300    0    1       1     1 132.040 53.646
#> 144 8813700  TRUE   0     1  300    0    1       1     1 118.690 59.691
#> 145 8813700  TRUE   0     1  300    0    1       1     1 111.830 53.775
#> 146 8813700  TRUE   0     1  300    0    1       1     1  95.780 54.516
#> 147 8813700  TRUE   0     1  300    0    1       1     1 135.530 51.307
#> 148 8813700  TRUE   0     1  300    0    1       1     1  99.496 51.160
#> 149 8813700  TRUE   0     1  300    0    1       1     1  89.308 55.116
#> 150 8813700  TRUE   0     1  300    0    1       1     1  98.103 57.835
#>     TIMEUNIT TRTACT         flag trtact   model
#> 1      Hours   3 mg Analysis set   3 mg xgxr001
#> 2      Hours   3 mg Analysis set   3 mg xgxr001
#> 3      Hours   3 mg Analysis set   3 mg xgxr001
#> 4      Hours   3 mg Analysis set   3 mg xgxr001
#> 5      Hours   3 mg Analysis set   3 mg xgxr001
#> 6      Hours   3 mg Analysis set   3 mg xgxr001
#> 7      Hours   3 mg Analysis set   3 mg xgxr001
#> 8      Hours   3 mg Analysis set   3 mg xgxr001
#> 9      Hours   3 mg Analysis set   3 mg xgxr001
#> 10     Hours   3 mg Analysis set   3 mg xgxr001
#> 11     Hours   3 mg Analysis set   3 mg xgxr001
#> 12     Hours   3 mg Analysis set   3 mg xgxr001
#> 13     Hours   3 mg Analysis set   3 mg xgxr001
#> 14     Hours   3 mg Analysis set   3 mg xgxr001
#> 15     Hours   3 mg Analysis set   3 mg xgxr001
#> 16     Hours   3 mg Analysis set   3 mg xgxr001
#> 17     Hours   3 mg Analysis set   3 mg xgxr001
#> 18     Hours   3 mg Analysis set   3 mg xgxr001
#> 19     Hours   3 mg Analysis set   3 mg xgxr001
#> 20     Hours   3 mg Analysis set   3 mg xgxr001
#> 21     Hours   3 mg Analysis set   3 mg xgxr001
#> 22     Hours   3 mg Analysis set   3 mg xgxr001
#> 23     Hours   3 mg Analysis set   3 mg xgxr001
#> 24     Hours   3 mg Analysis set   3 mg xgxr001
#> 25     Hours   3 mg Analysis set   3 mg xgxr001
#> 26     Hours   3 mg Analysis set   3 mg xgxr001
#> 27     Hours   3 mg Analysis set   3 mg xgxr001
#> 28     Hours   3 mg Analysis set   3 mg xgxr001
#> 29     Hours   3 mg Analysis set   3 mg xgxr001
#> 30     Hours   3 mg Analysis set   3 mg xgxr001
#> 31     Hours  10 mg Analysis set  10 mg xgxr001
#> 32     Hours  10 mg Analysis set  10 mg xgxr001
#> 33     Hours  10 mg Analysis set  10 mg xgxr001
#> 34     Hours  10 mg Analysis set  10 mg xgxr001
#> 35     Hours  10 mg Analysis set  10 mg xgxr001
#> 36     Hours  10 mg Analysis set  10 mg xgxr001
#> 37     Hours  10 mg Analysis set  10 mg xgxr001
#> 38     Hours  10 mg Analysis set  10 mg xgxr001
#> 39     Hours  10 mg Analysis set  10 mg xgxr001
#> 40     Hours  10 mg Analysis set  10 mg xgxr001
#> 41     Hours  10 mg Analysis set  10 mg xgxr001
#> 42     Hours  10 mg Analysis set  10 mg xgxr001
#> 43     Hours  10 mg Analysis set  10 mg xgxr001
#> 44     Hours  10 mg Analysis set  10 mg xgxr001
#> 45     Hours  10 mg Analysis set  10 mg xgxr001
#> 46     Hours  10 mg Analysis set  10 mg xgxr001
#> 47     Hours  10 mg Analysis set  10 mg xgxr001
#> 48     Hours  10 mg Analysis set  10 mg xgxr001
#> 49     Hours  10 mg Analysis set  10 mg xgxr001
#> 50     Hours  10 mg Analysis set  10 mg xgxr001
#> 51     Hours  10 mg Analysis set  10 mg xgxr001
#> 52     Hours  10 mg Analysis set  10 mg xgxr001
#> 53     Hours  10 mg Analysis set  10 mg xgxr001
#> 54     Hours  10 mg Analysis set  10 mg xgxr001
#> 55     Hours  10 mg Analysis set  10 mg xgxr001
#> 56     Hours  10 mg Analysis set  10 mg xgxr001
#> 57     Hours  10 mg Analysis set  10 mg xgxr001
#> 58     Hours  10 mg Analysis set  10 mg xgxr001
#> 59     Hours  10 mg Analysis set  10 mg xgxr001
#> 60     Hours  10 mg Analysis set  10 mg xgxr001
#> 61     Hours  30 mg Analysis set  30 mg xgxr001
#> 62     Hours  30 mg Analysis set  30 mg xgxr001
#> 63     Hours  30 mg Analysis set  30 mg xgxr001
#> 64     Hours  30 mg Analysis set  30 mg xgxr001
#> 65     Hours  30 mg Analysis set  30 mg xgxr001
#> 66     Hours  30 mg Analysis set  30 mg xgxr001
#> 67     Hours  30 mg Analysis set  30 mg xgxr001
#> 68     Hours  30 mg Analysis set  30 mg xgxr001
#> 69     Hours  30 mg Analysis set  30 mg xgxr001
#> 70     Hours  30 mg Analysis set  30 mg xgxr001
#> 71     Hours  30 mg Analysis set  30 mg xgxr001
#> 72     Hours  30 mg Analysis set  30 mg xgxr001
#> 73     Hours  30 mg Analysis set  30 mg xgxr001
#> 74     Hours  30 mg Analysis set  30 mg xgxr001
#> 75     Hours  30 mg Analysis set  30 mg xgxr001
#> 76     Hours  30 mg Analysis set  30 mg xgxr001
#> 77     Hours  30 mg Analysis set  30 mg xgxr001
#> 78     Hours  30 mg Analysis set  30 mg xgxr001
#> 79     Hours  30 mg Analysis set  30 mg xgxr001
#> 80     Hours  30 mg Analysis set  30 mg xgxr001
#> 81     Hours  30 mg Analysis set  30 mg xgxr001
#> 82     Hours  30 mg Analysis set  30 mg xgxr001
#> 83     Hours  30 mg Analysis set  30 mg xgxr001
#> 84     Hours  30 mg Analysis set  30 mg xgxr001
#> 85     Hours  30 mg Analysis set  30 mg xgxr001
#> 86     Hours  30 mg Analysis set  30 mg xgxr001
#> 87     Hours  30 mg Analysis set  30 mg xgxr001
#> 88     Hours  30 mg Analysis set  30 mg xgxr001
#> 89     Hours  30 mg Analysis set  30 mg xgxr001
#> 90     Hours  30 mg Analysis set  30 mg xgxr001
#> 91     Hours 100 mg Analysis set 100 mg xgxr001
#> 92     Hours 100 mg Analysis set 100 mg xgxr001
#> 93     Hours 100 mg Analysis set 100 mg xgxr001
#> 94     Hours 100 mg Analysis set 100 mg xgxr001
#> 95     Hours 100 mg Analysis set 100 mg xgxr001
#> 96     Hours 100 mg Analysis set 100 mg xgxr001
#> 97     Hours 100 mg Analysis set 100 mg xgxr001
#> 98     Hours 100 mg Analysis set 100 mg xgxr001
#> 99     Hours 100 mg Analysis set 100 mg xgxr001
#> 100    Hours 100 mg Analysis set 100 mg xgxr001
#> 101    Hours 100 mg Analysis set 100 mg xgxr001
#> 102    Hours 100 mg Analysis set 100 mg xgxr001
#> 103    Hours 100 mg Analysis set 100 mg xgxr001
#> 104    Hours 100 mg Analysis set 100 mg xgxr001
#> 105    Hours 100 mg Analysis set 100 mg xgxr001
#> 106    Hours 100 mg Analysis set 100 mg xgxr001
#> 107    Hours 100 mg Analysis set 100 mg xgxr001
#> 108    Hours 100 mg Analysis set 100 mg xgxr001
#> 109    Hours 100 mg Analysis set 100 mg xgxr001
#> 110    Hours 100 mg Analysis set 100 mg xgxr001
#> 111    Hours 100 mg Analysis set 100 mg xgxr001
#> 112    Hours 100 mg Analysis set 100 mg xgxr001
#> 113    Hours 100 mg Analysis set 100 mg xgxr001
#> 114    Hours 100 mg Analysis set 100 mg xgxr001
#> 115    Hours 100 mg Analysis set 100 mg xgxr001
#> 116    Hours 100 mg Analysis set 100 mg xgxr001
#> 117    Hours 100 mg Analysis set 100 mg xgxr001
#> 118    Hours 100 mg Analysis set 100 mg xgxr001
#> 119    Hours 100 mg Analysis set 100 mg xgxr001
#> 120    Hours 100 mg Analysis set 100 mg xgxr001
#> 121    Hours 300 mg Analysis set 300 mg xgxr001
#> 122    Hours 300 mg Analysis set 300 mg xgxr001
#> 123    Hours 300 mg Analysis set 300 mg xgxr001
#> 124    Hours 300 mg Analysis set 300 mg xgxr001
#> 125    Hours 300 mg Analysis set 300 mg xgxr001
#> 126    Hours 300 mg Analysis set 300 mg xgxr001
#> 127    Hours 300 mg Analysis set 300 mg xgxr001
#> 128    Hours 300 mg Analysis set 300 mg xgxr001
#> 129    Hours 300 mg Analysis set 300 mg xgxr001
#> 130    Hours 300 mg Analysis set 300 mg xgxr001
#> 131    Hours 300 mg Analysis set 300 mg xgxr001
#> 132    Hours 300 mg Analysis set 300 mg xgxr001
#> 133    Hours 300 mg Analysis set 300 mg xgxr001
#> 134    Hours 300 mg Analysis set 300 mg xgxr001
#> 135    Hours 300 mg Analysis set 300 mg xgxr001
#> 136    Hours 300 mg Analysis set 300 mg xgxr001
#> 137    Hours 300 mg Analysis set 300 mg xgxr001
#> 138    Hours 300 mg Analysis set 300 mg xgxr001
#> 139    Hours 300 mg Analysis set 300 mg xgxr001
#> 140    Hours 300 mg Analysis set 300 mg xgxr001
#> 141    Hours 300 mg Analysis set 300 mg xgxr001
#> 142    Hours 300 mg Analysis set 300 mg xgxr001
#> 143    Hours 300 mg Analysis set 300 mg xgxr001
#> 144    Hours 300 mg Analysis set 300 mg xgxr001
#> 145    Hours 300 mg Analysis set 300 mg xgxr001
#> 146    Hours 300 mg Analysis set 300 mg xgxr001
#> 147    Hours 300 mg Analysis set 300 mg xgxr001
#> 148    Hours 300 mg Analysis set 300 mg xgxr001
#> 149    Hours 300 mg Analysis set 300 mg xgxr001
#> 150    Hours 300 mg Analysis set 300 mg xgxr001

### Without an ID column we get non-varying columns
findCovs(dat)

#>      TVKA  TVV2    TVCL    TVV3     TVQ      V3       Q nmout BLQ CYCLE FLAG
#> 1 0.18116 0.042 0.72491 0.17857 8813700 0.17857 8813700  TRUE   0     1    0
#>   PART PROFDAY STUDY TIMEUNIT         flag   model
#> 1    1       1     1    Hours Analysis set xgxr001