Extract columns that do not vary within variables in a data.frame

Extract columns that do not vary within variables in a data.frame

findCovs(data, cols.id = NULL, as.fun = NULL)

Arguments

data
data.frame in which to look for covariates

cols.id
covariates will be searched for in combinations of values in these columns. Often cols.id will be either empty or ID. But it can also be both say c("ID","DRUG") or c("ID","TRT").

as.fun
The default is to return data in data.tables. Pass a function in as.fun to convert to something else. If data.frames are wanted, use as.fun=as.data.frame.

See also

Examples

dat <- NMscanData(NMdata_filepath("examples/nonmem/xgxr001.lst"))

#> Number of output tables read: 1.

#> Read delimited text input data file.

#> 
#> Input data columns will be appended to output data. However, column(s) were identified as unique identifiers, present in both input and output data. If this column or one of these columns is not modified by the Nonmem run, consider using this in col.row for a robust merge of input and output data. Candidate columns: ROW

#> To skip this check, please specify either col.row (recommended) or cbind.by.filters.

### very common use
findCovs(dat,cols.id="ID")

#>       ID    TVKA  TVV2    TVCL    TVV3     TVQ      KA       V2      CL      V3
#>   1:  31 0.18116 0.042 0.72491 0.17857 8813700 0.18116 0.042000 0.72491 0.17857
#>   2:  32 0.18116 0.042 0.72491 0.17857 8813700 0.18116 0.042000 0.72491 0.17857
#>   3:  33 0.18116 0.042 0.72491 0.17857 8813700 0.18116 0.042000 0.72491 0.17857
#>   4:  34 0.18116 0.042 0.72491 0.17857 8813700 0.18116 0.042000 0.72491 0.17857
#>   5:  35 0.18116 0.042 0.72491 0.17857 8813700 0.20098 0.044867 0.72487 0.17857
#>  ---                                                                           
#> 146: 176 0.18116 0.042 0.72491 0.17857 8813700 0.11236 0.058950 0.72475 0.17857
#> 147: 177 0.18116 0.042 0.72491 0.17857 8813700 0.13466 0.041848 0.72495 0.17857
#> 148: 178 0.18116 0.042 0.72491 0.17857 8813700 0.15852 0.051512 0.72486 0.17857
#> 149: 179 0.18116 0.042 0.72491 0.17857 8813700 0.14838 0.067208 0.72512 0.17857
#> 150: 180 0.18116 0.042 0.72491 0.17857 8813700 0.22131 0.022291 0.72520 0.17857
#>            Q nmout BLQ CYCLE DOSE FLAG PART PROFDAY STUDY WEIGHTB   eff0
#>   1: 8813700  TRUE   0     1    3    0    1       1     1  87.031 56.461
#>   2: 8813700  TRUE   0     1    3    0    1       1     1 100.620 45.096
#>   3: 8813700  TRUE   0     1    3    0    1       1     1 112.990 49.189
#>   4: 8813700  TRUE   0     1    3    0    1       1     1 144.130 56.239
#>   5: 8813700  TRUE   0     1    3    0    1       1     1  80.384 58.419
#>  ---                                                                    
#> 146: 8813700  TRUE   0     1  300    0    1       1     1  95.780 54.516
#> 147: 8813700  TRUE   0     1  300    0    1       1     1 135.530 51.307
#> 148: 8813700  TRUE   0     1  300    0    1       1     1  99.496 51.160
#> 149: 8813700  TRUE   0     1  300    0    1       1     1  89.308 55.116
#> 150: 8813700  TRUE   0     1  300    0    1       1     1  98.103 57.835
#>      TIMEUNIT TRTACT         flag trtact   model
#>   1:    Hours   3 mg Analysis set   3 mg xgxr001
#>   2:    Hours   3 mg Analysis set   3 mg xgxr001
#>   3:    Hours   3 mg Analysis set   3 mg xgxr001
#>   4:    Hours   3 mg Analysis set   3 mg xgxr001
#>   5:    Hours   3 mg Analysis set   3 mg xgxr001
#>  ---                                            
#> 146:    Hours 300 mg Analysis set 300 mg xgxr001
#> 147:    Hours 300 mg Analysis set 300 mg xgxr001
#> 148:    Hours 300 mg Analysis set 300 mg xgxr001
#> 149:    Hours 300 mg Analysis set 300 mg xgxr001
#> 150:    Hours 300 mg Analysis set 300 mg xgxr001

### Without an ID column we get non-varying columns
findCovs(dat)

#>       TVKA  TVV2    TVCL    TVV3     TVQ      V3       Q nmout BLQ CYCLE FLAG
#> 1: 0.18116 0.042 0.72491 0.17857 8813700 0.17857 8813700  TRUE   0     1    0
#>    PART PROFDAY STUDY TIMEUNIT         flag   model
#> 1:    1       1     1    Hours Analysis set xgxr001