Extract columns that vary within values of other columns

This function provides an automated method to extract covariate-like columns. The user decides which columns these variables cannot vary within. So if you have repeated measures for each ID, this function can find the columns that are constant within ID and their unique values for each ID. Or, you can provide a combination of id.cols, say ID and STUDY, and get variables that do not vary within unique combinations of these.

findCovs(data, by = NULL, cols.id, as.fun = NULL)

Arguments

data

data.frame in which to look for covariates

by

covariates will be searched for in combinations of values in these columns. Often by will be either empty or ID. But it can also be both say c("ID","DRUG") or c("ID","TRT").

cols.id

Deprecated. Use by instead.

as.fun

The default is to return a data.table if data is a data.table and return a data.frame in all other cases. Pass a function in as.fun to convert to something else. If data is not a data.table, the default can be configured using NMdataConf.

Value

a data set with one observation per combination of values of variables listed in by.

See also

Examples

dt1=data.frame(ID=c(1,1,2,2),
               OCC=c(1,2,1,2),
               ## ID level
               eta1=c(1,1,3,3),
               ## occasion level
               eta2=c(1,3,1,5),
               ## not used
               eta3=0
               )
## model level
findCovs(dt1)
#>   eta3
#> 1    0
## ID level
findCovs(dt1,"ID")
#>   ID eta1 eta3
#> 1  1    1    0
#> 2  2    3    0
## acual ID level
findVars(findCovs(dt1,"ID"))
#>   ID eta1
#> 1  1    1
#> 2  2    3
## occasion level
findCovs(findVars(dt1,"ID"),c("ID","OCC"))
#>   ID OCC eta2
#> 1  1   1    1
#> 2  1   2    3
#> 3  2   1    1
#> 4  2   2    5
## Based on a "real data example"
dat <- NMscanData(system.file("examples/nonmem/xgxr001.lst", package = "NMdata"))
#> Model:  xgxr001 
#> 
#> Used tables, contents shown as used/total:
#>               file     rows columns     IDs
#>    xgxr001_res.txt  905/905   16/16 150/150
#>  xgxr1.csv (input) 905/1502   22/24 150/150
#>           (result)      905    38+2     150
#> 
#> Input and output data merged by: ROW 
#> 
#> Distribution of rows on event types in returned data:
#>  EVID CMT output result
#>     0   2    755    755
#>     1   1    150    150
#>   All All    905    905
findCovs(dat,by="ID")
#>      ID FLAG STUDY   TVKA  TVV2    TVCL    TVV3     TVQ       KA       V2
#> 1    31    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 2    32    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 3    33    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 4    34    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 5    35    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.200980 0.044864
#> 6    36    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 7    37    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 8    38    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 9    39    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 10   40    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.198940 0.044706
#> 11   41    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.197620 0.044184
#> 12   42    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 13   43    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.210440 0.046147
#> 14   44    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 15   45    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 16   46    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 17   47    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.193720 0.044403
#> 18   48    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.193430 0.044556
#> 19   49    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 20   50    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.199410 0.044394
#> 21   51    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.232850 0.051847
#> 22   52    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 23   53    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 24   54    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.197720 0.045294
#> 25   55    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.196600 0.045038
#> 26   56    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 27   57    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 28   58    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 29   59    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.235940 0.052063
#> 30   60    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181200 0.042000
#> 31   61    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.226410 0.050266
#> 32   62    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.182770 0.042690
#> 33   63    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.157830 0.038184
#> 34   64    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.167120 0.040140
#> 35   65    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.175540 0.041745
#> 36   66    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.156180 0.038454
#> 37   67    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.154850 0.037744
#> 38   68    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.207800 0.045034
#> 39   69    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.200890 0.043796
#> 40   70    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.177060 0.041878
#> 41   71    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.196970 0.044786
#> 42   72    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.193820 0.048997
#> 43   73    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.192730 0.049952
#> 44   74    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.237960 0.053130
#> 45   75    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.224180 0.051662
#> 46   76    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.184770 0.042518
#> 47   77    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.243730 0.059378
#> 48   78    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.154990 0.038185
#> 49   79    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.183170 0.040628
#> 50   80    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.168150 0.040644
#> 51   81    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.306230 0.059140
#> 52   82    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.164380 0.039367
#> 53   83    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.217320 0.049514
#> 54   84    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.195520 0.042983
#> 55   85    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.242240 0.053765
#> 56   86    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.198600 0.046308
#> 57   87    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.202930 0.047993
#> 58   88    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.196620 0.046509
#> 59   89    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.180480 0.041716
#> 60   90    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.176060 0.041935
#> 61   91    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.303930 0.058836
#> 62   92    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.135850 0.041090
#> 63   93    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.196310 0.032508
#> 64   94    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.152410 0.035156
#> 65   95    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.160700 0.040610
#> 66   96    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.159810 0.038137
#> 67   97    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.164970 0.033911
#> 68   98    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.111970 0.063107
#> 69   99    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.191590 0.045655
#> 70  100    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.139780 0.032777
#> 71  101    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.218420 0.038330
#> 72  102    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.284330 0.044044
#> 73  103    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.111050 0.042953
#> 74  104    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.146970 0.037793
#> 75  105    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.311330 0.048329
#> 76  106    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.140460 0.035070
#> 77  107    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.192870 0.035889
#> 78  108    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.150810 0.036928
#> 79  109    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.122370 0.045333
#> 80  110    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.205090 0.036557
#> 81  111    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.297940 0.041776
#> 82  112    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.235920 0.043718
#> 83  113    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.200610 0.050441
#> 84  114    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.163240 0.044282
#> 85  115    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.191200 0.035260
#> 86  116    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.141670 0.033432
#> 87  117    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.107480 0.030168
#> 88  118    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.147590 0.031947
#> 89  119    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.196130 0.034621
#> 90  120    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.117490 0.032430
#> 91  121    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.220570 0.038709
#> 92  122    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.136510 0.037994
#> 93  123    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.160980 0.042470
#> 94  124    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.219570 0.033761
#> 95  125    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.153730 0.052726
#> 96  126    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.295070 0.042096
#> 97  127    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.075592 0.086623
#> 98  128    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.139970 0.039725
#> 99  129    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.204480 0.044707
#> 100 130    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.189990 0.042004
#> 101 131    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.084695 0.065501
#> 102 132    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.103510 0.069172
#> 103 133    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.205720 0.032188
#> 104 134    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.614130 0.036348
#> 105 135    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.302290 0.041085
#> 106 136    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.090525 0.041405
#> 107 137    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.331460 0.031894
#> 108 138    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.119470 0.037073
#> 109 139    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.091157 0.075868
#> 110 140    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.308030 0.040073
#> 111 141    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.088011 0.047181
#> 112 142    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.214470 0.028516
#> 113 143    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.110100 0.063143
#> 114 144    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.142510 0.058877
#> 115 145    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.221800 0.046486
#> 116 146    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.181210 0.051057
#> 117 147    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.167840 0.036615
#> 118 148    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.144060 0.048420
#> 119 149    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.329890 0.031505
#> 120 150    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.240510 0.026796
#> 121 151    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.074444 0.092298
#> 122 152    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.116820 0.051434
#> 123 153    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.422730 0.034489
#> 124 154    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.096624 0.051236
#> 125 155    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.151840 0.035453
#> 126 156    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.370260 0.022135
#> 127 157    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.170190 0.066894
#> 128 158    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.125970 0.056376
#> 129 159    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.133710 0.047449
#> 130 160    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.406220 0.020625
#> 131 161    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.144900 0.069480
#> 132 162    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.250600 0.020414
#> 133 163    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.192560 0.024113
#> 134 164    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.123500 0.061047
#> 135 165    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.401710 0.020531
#> 136 166    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.174450 0.038936
#> 137 167    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.282000 0.033686
#> 138 168    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.285010 0.040010
#> 139 169    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.116410 0.051361
#> 140 170    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.297080 0.021852
#> 141 171    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.091625 0.054324
#> 142 172    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.355210 0.022146
#> 143 173    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.324170 0.038394
#> 144 174    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.135430 0.057318
#> 145 175    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.312240 0.036860
#> 146 176    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.112360 0.058919
#> 147 177    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.134660 0.041833
#> 148 178    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.158520 0.051490
#> 149 179    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.148380 0.067188
#> 150 180    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.221270 0.022290
#>          CL      V3       Q BLQ CYCLE DOSE PART PROFDAY WEIGHTB   eff0 TIMEUNIT
#> 1   0.72457 0.17848 2307400   0     1    3    1       1  87.031 56.461    Hours
#> 2   0.72457 0.17848 2307400   0     1    3    1       1 100.620 45.096    Hours
#> 3   0.72457 0.17848 2307400   0     1    3    1       1 112.990 49.189    Hours
#> 4   0.72457 0.17848 2307400   0     1    3    1       1 144.130 56.239    Hours
#> 5   0.72451 0.17848 2307400   0     1    3    1       1  80.384 58.419    Hours
#> 6   0.72457 0.17848 2307400   0     1    3    1       1 120.860 58.163    Hours
#> 7   0.72457 0.17848 2307400   0     1    3    1       1 138.440 42.122    Hours
#> 8   0.72457 0.17848 2307400   0     1    3    1       1 123.350 59.180    Hours
#> 9   0.72457 0.17848 2307400   0     1    3    1       1 137.080 47.222    Hours
#> 10  0.72451 0.17848 2307400   0     1    3    1       1 134.090 49.498    Hours
#> 11  0.72453 0.17848 2307400   0     1    3    1       1  80.018 41.362    Hours
#> 12  0.72457 0.17848 2307400   0     1    3    1       1 142.900 54.624    Hours
#> 13  0.72449 0.17848 2307400   0     1    3    1       1  97.637 51.154    Hours
#> 14  0.72457 0.17848 2307400   0     1    3    1       1 105.310 54.303    Hours
#> 15  0.72457 0.17848 2307400   0     1    3    1       1  83.345 59.970    Hours
#> 16  0.72457 0.17848 2307400   0     1    3    1       1 148.540 40.533    Hours
#> 17  0.72451 0.17848 2307400   0     1    3    1       1 121.060 42.542    Hours
#> 18  0.72451 0.17848 2307400   0     1    3    1       1  97.761 51.290    Hours
#> 19  0.72457 0.17848 2307400   0     1    3    1       1 144.240 45.758    Hours
#> 20  0.72453 0.17848 2307400   0     1    3    1       1 111.400 44.106    Hours
#> 21  0.72436 0.17848 2307400   0     1    3    1       1 144.870 42.832    Hours
#> 22  0.72457 0.17848 2307400   0     1    3    1       1 136.980 48.142    Hours
#> 23  0.72457 0.17848 2307400   0     1    3    1       1 124.950 52.037    Hours
#> 24  0.72449 0.17848 2307400   0     1    3    1       1 139.260 54.637    Hours
#> 25  0.72450 0.17848 2307400   0     1    3    1       1 127.350 57.822    Hours
#> 26  0.72457 0.17848 2307400   0     1    3    1       1 136.260 48.192    Hours
#> 27  0.72457 0.17848 2307400   0     1    3    1       1 139.740 45.235    Hours
#> 28  0.72457 0.17848 2307400   0     1    3    1       1  90.185 51.918    Hours
#> 29  0.72435 0.17848 2307400   0     1    3    1       1 142.720 54.874    Hours
#> 30  0.72457 0.17848 2307400   0     1    3    1       1  91.702 57.376    Hours
#> 31  0.72437 0.17848 2307400   0     1   10    1       1 133.040 58.744    Hours
#> 32  0.72455 0.17848 2307400   0     1   10    1       1 104.290 40.081    Hours
#> 33  0.72466 0.17848 2307400   0     1   10    1       1 107.650 49.956    Hours
#> 34  0.72462 0.17848 2307400   0     1   10    1       1 141.420 56.587    Hours
#> 35  0.72457 0.17848 2307400   0     1   10    1       1 113.970 50.127    Hours
#> 36  0.72464 0.17848 2307400   0     1   10    1       1 112.190 49.280    Hours
#> 37  0.72467 0.17848 2307400   0     1   10    1       1  87.380 48.628    Hours
#> 38  0.72451 0.17848 2307400   0     1   10    1       1 126.680 47.701    Hours
#> 39  0.72459 0.17848 2307400   0     1   10    1       1 110.810 44.421    Hours
#> 40  0.72455 0.17848 2307400   0     1   10    1       1 142.900 50.431    Hours
#> 41  0.72452 0.17848 2307400   0     1   10    1       1  88.158 58.193    Hours
#> 42  0.72440 0.17848 2307400   0     1   10    1       1 115.390 59.961    Hours
#> 43  0.72435 0.17848 2307400   0     1   10    1       1  91.806 41.836    Hours
#> 44  0.72430 0.17848 2307400   0     1   10    1       1 111.080 42.848    Hours
#> 45  0.72431 0.17848 2307400   0     1   10    1       1 120.870 56.245    Hours
#> 46  0.72456 0.17848 2307400   0     1   10    1       1 102.040 47.031    Hours
#> 47  0.72416 0.17848 2307400   0     1   10    1       1 143.750 52.876    Hours
#> 48  0.72465 0.17848 2307400   0     1   10    1       1 148.760 46.962    Hours
#> 49  0.72465 0.17848 2307400   0     1   10    1       1 128.350 43.855    Hours
#> 50  0.72457 0.17848 2307400   0     1   10    1       1 143.310 57.924    Hours
#> 51  0.72432 0.17848 2307400   0     1   10    1       1 145.680 48.497    Hours
#> 52  0.72462 0.17848 2307400   0     1   10    1       1 125.550 48.890    Hours
#> 53  0.72439 0.17848 2307400   0     1   10    1       1 124.870 41.692    Hours
#> 54  0.72456 0.17848 2307400   0     1   10    1       1 127.300 47.241    Hours
#> 55  0.72433 0.17848 2307400   0     1   10    1       1 139.600 55.084    Hours
#> 56  0.72443 0.17848 2307400   0     1   10    1       1 147.170 42.582    Hours
#> 57  0.72441 0.17848 2307400   0     1   10    1       1 112.100 59.316    Hours
#> 58  0.72445 0.17848 2307400   0     1   10    1       1 122.970 49.694    Hours
#> 59  0.72458 0.17848 2307400   0     1   10    1       1 101.600 40.517    Hours
#> 60  0.72454 0.17848 2307400   0     1   10    1       1 126.330 46.436    Hours
#> 61  0.72415 0.17848 2307400   0     1   30    1       1  91.716 58.817    Hours
#> 62  0.72446 0.17848 2307400   0     1   30    1       1 101.980 41.972    Hours
#> 63  0.72500 0.17848 2307400   0     1   30    1       1 131.860 57.696    Hours
#> 64  0.72467 0.17848 2307400   0     1   30    1       1  84.754 42.902    Hours
#> 65  0.72448 0.17848 2307400   0     1   30    1       1  83.878 41.517    Hours
#> 66  0.72451 0.17848 2307400   0     1   30    1       1 121.110 57.137    Hours
#> 67  0.72480 0.17848 2307400   0     1   30    1       1 133.690 55.847    Hours
#> 68  0.72410 0.17848 2307400   0     1   30    1       1  80.206 55.302    Hours
#> 69  0.72439 0.17848 2307400   0     1   30    1       1  96.745 49.018    Hours
#> 70  0.72480 0.17848 2307400   0     1   30    1       1 109.040 46.163    Hours
#> 71  0.72471 0.17848 2307400   0     1   30    1       1 142.040 53.020    Hours
#> 72  0.72456 0.17848 2307400   0     1   30    1       1 111.700 53.996    Hours
#> 73  0.72456 0.17848 2307400   0     1   30    1       1  92.732 50.635    Hours
#> 74  0.72465 0.17848 2307400   0     1   30    1       1  83.444 58.264    Hours
#> 75  0.72452 0.17848 2307400   0     1   30    1       1 121.690 49.004    Hours
#> 76  0.72477 0.17848 2307400   0     1   30    1       1 133.300 53.440    Hours
#> 77  0.72479 0.17848 2307400   0     1   30    1       1  84.755 49.630    Hours
#> 78  0.72476 0.17848 2307400   0     1   30    1       1  95.846 46.434    Hours
#> 79  0.72447 0.17848 2307400   0     1   30    1       1 113.810 57.663    Hours
#> 80  0.72486 0.17848 2307400   0     1   30    1       1 116.780 58.513    Hours
#> 81  0.72462 0.17848 2307400   0     1   30    1       1 138.870 58.687    Hours
#> 82  0.72450 0.17848 2307400   0     1   30    1       1 114.000 50.922    Hours
#> 83  0.72430 0.17848 2307400   0     1   30    1       1 114.260 49.663    Hours
#> 84  0.72434 0.17848 2307400   0     1   30    1       1 105.090 49.476    Hours
#> 85  0.72472 0.17848 2307400   0     1   30    1       1 129.360 50.889    Hours
#> 86  0.72475 0.17848 2307400   0     1   30    1       1 109.160 42.200    Hours
#> 87  0.72488 0.17848 2307400   0     1   30    1       1 116.960 45.204    Hours
#> 88  0.72489 0.17848 2307400   0     1   30    1       1 119.330 40.657    Hours
#> 89  0.72478 0.17848 2307400   0     1   30    1       1 130.040 57.173    Hours
#> 90  0.72483 0.17848 2307400   0     1   30    1       1  95.384 47.696    Hours
#> 91  0.72449 0.17848 2307400   0     1  100    1       1 136.770 47.184    Hours
#> 92  0.72440 0.17848 2307400   0     1  100    1       1  94.357 49.776    Hours
#> 93  0.72453 0.17848 2307400   0     1  100    1       1  81.491 57.013    Hours
#> 94  0.72473 0.17848 2307400   0     1  100    1       1 115.370 51.224    Hours
#> 95  0.72442 0.17848 2307400   0     1  100    1       1 148.110 41.539    Hours
#> 96  0.72464 0.17848 2307400   0     1  100    1       1 148.230 48.961    Hours
#> 97  0.72361 0.17848 2307400   0     1  100    1       1 133.160 42.732    Hours
#> 98  0.72492 0.17848 2307400   0     1  100    1       1  82.140 49.704    Hours
#> 99  0.72442 0.17848 2307400   0     1  100    1       1  88.087 46.169    Hours
#> 100 0.72444 0.17848 2307400   0     1  100    1       1 142.170 49.395    Hours
#> 101 0.72403 0.17848 2307400   0     1  100    1       1 117.810 57.872    Hours
#> 102 0.72405 0.17848 2307400   0     1  100    1       1 132.710 42.860    Hours
#> 103 0.72480 0.17848 2307400   0     1  100    1       1 114.950 40.940    Hours
#> 104 0.72486 0.17848 2307400   0     1  100    1       1 145.100 54.023    Hours
#> 105 0.72444 0.17848 2307400   0     1  100    1       1 117.550 55.676    Hours
#> 106 0.72437 0.17848 2307400   0     1  100    1       1 117.820 45.134    Hours
#> 107 0.72504 0.17848 2307400   0     1  100    1       1 133.540 45.305    Hours
#> 108 0.72463 0.17848 2307400   0     1  100    1       1 119.450 49.917    Hours
#> 109 0.72387 0.17848 2307400   0     1  100    1       1 149.550 43.930    Hours
#> 110 0.72440 0.17848 2307400   0     1  100    1       1 112.980 48.596    Hours
#> 111 0.72417 0.17848 2307400   0     1  100    1       1 130.200 44.755    Hours
#> 112 0.72511 0.17848 2307400   0     1  100    1       1 147.080 50.895    Hours
#> 113 0.72416 0.17848 2307400   0     1  100    1       1  96.265 43.053    Hours
#> 114 0.72422 0.17848 2307400   0     1  100    1       1 132.650 51.302    Hours
#> 115 0.72468 0.17848 2307400   0     1  100    1       1 118.270 50.193    Hours
#> 116 0.72453 0.17848 2307400   0     1  100    1       1 114.600 56.089    Hours
#> 117 0.72449 0.17848 2307400   0     1  100    1       1 141.930 54.312    Hours
#> 118 0.72440 0.17848 2307400   0     1  100    1       1  99.604 44.034    Hours
#> 119 0.72491 0.17848 2307400   0     1  100    1       1 111.110 55.695    Hours
#> 120 0.72487 0.17848 2307400   0     1  100    1       1 111.820 55.035    Hours
#> 121 0.72344 0.17848 2307400   0     1  300    1       1  88.319 53.035    Hours
#> 122 0.72418 0.17848 2307400   0     1  300    1       1 121.320 47.391    Hours
#> 123 0.72475 0.17848 2307400   0     1  300    1       1  93.475 56.259    Hours
#> 124 0.72463 0.17848 2307400   0     1  300    1       1 147.680 52.824    Hours
#> 125 0.72470 0.17848 2307400   0     1  300    1       1  96.953 58.300    Hours
#> 126 0.72486 0.17848 2307400   0     1  300    1       1 140.810 46.160    Hours
#> 127 0.72452 0.17848 2307400   0     1  300    1       1  94.526 53.894    Hours
#> 128 0.72400 0.17848 2307400   0     1  300    1       1 146.820 54.044    Hours
#> 129 0.72431 0.17848 2307400   0     1  300    1       1 145.530 55.699    Hours
#> 130 0.72509 0.17848 2307400   0     1  300    1       1 127.400 50.455    Hours
#> 131 0.72511 0.17848 2307400   0     1  300    1       1 101.580 55.400    Hours
#> 132 0.72516 0.17848 2307400   0     1  300    1       1 102.620 59.111    Hours
#> 133 0.72503 0.17848 2307400   0     1  300    1       1  85.221 41.015    Hours
#> 134 0.72420 0.17848 2307400   0     1  300    1       1  84.974 54.064    Hours
#> 135 0.72526 0.17848 2307400   0     1  300    1       1 124.230 53.337    Hours
#> 136 0.72467 0.17848 2307400   0     1  300    1       1 101.180 59.792    Hours
#> 137 0.72484 0.17848 2307400   0     1  300    1       1 119.050 45.097    Hours
#> 138 0.72464 0.17848 2307400   0     1  300    1       1  95.155 46.504    Hours
#> 139 0.72411 0.17848 2307400   0     1  300    1       1 124.980 54.155    Hours
#> 140 0.72535 0.17848 2307400   0     1  300    1       1 102.150 51.862    Hours
#> 141 0.72422 0.17848 2307400   0     1  300    1       1 136.230 47.295    Hours
#> 142 0.72534 0.17848 2307400   0     1  300    1       1  80.888 49.242    Hours
#> 143 0.72502 0.17848 2307400   0     1  300    1       1 132.040 53.646    Hours
#> 144 0.72415 0.17848 2307400   0     1  300    1       1 118.690 59.691    Hours
#> 145 0.72503 0.17848 2307400   0     1  300    1       1 111.830 53.775    Hours
#> 146 0.72431 0.17848 2307400   0     1  300    1       1  95.780 54.516    Hours
#> 147 0.72465 0.17848 2307400   0     1  300    1       1 135.530 51.307    Hours
#> 148 0.72449 0.17848 2307400   0     1  300    1       1  99.496 51.160    Hours
#> 149 0.72492 0.17848 2307400   0     1  300    1       1  89.308 55.116    Hours
#> 150 0.72506 0.17848 2307400   0     1  300    1       1  98.103 57.835    Hours
#>     TRTACT trtact   model nmout
#> 1     3 mg   3 mg xgxr001  TRUE
#> 2     3 mg   3 mg xgxr001  TRUE
#> 3     3 mg   3 mg xgxr001  TRUE
#> 4     3 mg   3 mg xgxr001  TRUE
#> 5     3 mg   3 mg xgxr001  TRUE
#> 6     3 mg   3 mg xgxr001  TRUE
#> 7     3 mg   3 mg xgxr001  TRUE
#> 8     3 mg   3 mg xgxr001  TRUE
#> 9     3 mg   3 mg xgxr001  TRUE
#> 10    3 mg   3 mg xgxr001  TRUE
#> 11    3 mg   3 mg xgxr001  TRUE
#> 12    3 mg   3 mg xgxr001  TRUE
#> 13    3 mg   3 mg xgxr001  TRUE
#> 14    3 mg   3 mg xgxr001  TRUE
#> 15    3 mg   3 mg xgxr001  TRUE
#> 16    3 mg   3 mg xgxr001  TRUE
#> 17    3 mg   3 mg xgxr001  TRUE
#> 18    3 mg   3 mg xgxr001  TRUE
#> 19    3 mg   3 mg xgxr001  TRUE
#> 20    3 mg   3 mg xgxr001  TRUE
#> 21    3 mg   3 mg xgxr001  TRUE
#> 22    3 mg   3 mg xgxr001  TRUE
#> 23    3 mg   3 mg xgxr001  TRUE
#> 24    3 mg   3 mg xgxr001  TRUE
#> 25    3 mg   3 mg xgxr001  TRUE
#> 26    3 mg   3 mg xgxr001  TRUE
#> 27    3 mg   3 mg xgxr001  TRUE
#> 28    3 mg   3 mg xgxr001  TRUE
#> 29    3 mg   3 mg xgxr001  TRUE
#> 30    3 mg   3 mg xgxr001  TRUE
#> 31   10 mg  10 mg xgxr001  TRUE
#> 32   10 mg  10 mg xgxr001  TRUE
#> 33   10 mg  10 mg xgxr001  TRUE
#> 34   10 mg  10 mg xgxr001  TRUE
#> 35   10 mg  10 mg xgxr001  TRUE
#> 36   10 mg  10 mg xgxr001  TRUE
#> 37   10 mg  10 mg xgxr001  TRUE
#> 38   10 mg  10 mg xgxr001  TRUE
#> 39   10 mg  10 mg xgxr001  TRUE
#> 40   10 mg  10 mg xgxr001  TRUE
#> 41   10 mg  10 mg xgxr001  TRUE
#> 42   10 mg  10 mg xgxr001  TRUE
#> 43   10 mg  10 mg xgxr001  TRUE
#> 44   10 mg  10 mg xgxr001  TRUE
#> 45   10 mg  10 mg xgxr001  TRUE
#> 46   10 mg  10 mg xgxr001  TRUE
#> 47   10 mg  10 mg xgxr001  TRUE
#> 48   10 mg  10 mg xgxr001  TRUE
#> 49   10 mg  10 mg xgxr001  TRUE
#> 50   10 mg  10 mg xgxr001  TRUE
#> 51   10 mg  10 mg xgxr001  TRUE
#> 52   10 mg  10 mg xgxr001  TRUE
#> 53   10 mg  10 mg xgxr001  TRUE
#> 54   10 mg  10 mg xgxr001  TRUE
#> 55   10 mg  10 mg xgxr001  TRUE
#> 56   10 mg  10 mg xgxr001  TRUE
#> 57   10 mg  10 mg xgxr001  TRUE
#> 58   10 mg  10 mg xgxr001  TRUE
#> 59   10 mg  10 mg xgxr001  TRUE
#> 60   10 mg  10 mg xgxr001  TRUE
#> 61   30 mg  30 mg xgxr001  TRUE
#> 62   30 mg  30 mg xgxr001  TRUE
#> 63   30 mg  30 mg xgxr001  TRUE
#> 64   30 mg  30 mg xgxr001  TRUE
#> 65   30 mg  30 mg xgxr001  TRUE
#> 66   30 mg  30 mg xgxr001  TRUE
#> 67   30 mg  30 mg xgxr001  TRUE
#> 68   30 mg  30 mg xgxr001  TRUE
#> 69   30 mg  30 mg xgxr001  TRUE
#> 70   30 mg  30 mg xgxr001  TRUE
#> 71   30 mg  30 mg xgxr001  TRUE
#> 72   30 mg  30 mg xgxr001  TRUE
#> 73   30 mg  30 mg xgxr001  TRUE
#> 74   30 mg  30 mg xgxr001  TRUE
#> 75   30 mg  30 mg xgxr001  TRUE
#> 76   30 mg  30 mg xgxr001  TRUE
#> 77   30 mg  30 mg xgxr001  TRUE
#> 78   30 mg  30 mg xgxr001  TRUE
#> 79   30 mg  30 mg xgxr001  TRUE
#> 80   30 mg  30 mg xgxr001  TRUE
#> 81   30 mg  30 mg xgxr001  TRUE
#> 82   30 mg  30 mg xgxr001  TRUE
#> 83   30 mg  30 mg xgxr001  TRUE
#> 84   30 mg  30 mg xgxr001  TRUE
#> 85   30 mg  30 mg xgxr001  TRUE
#> 86   30 mg  30 mg xgxr001  TRUE
#> 87   30 mg  30 mg xgxr001  TRUE
#> 88   30 mg  30 mg xgxr001  TRUE
#> 89   30 mg  30 mg xgxr001  TRUE
#> 90   30 mg  30 mg xgxr001  TRUE
#> 91  100 mg 100 mg xgxr001  TRUE
#> 92  100 mg 100 mg xgxr001  TRUE
#> 93  100 mg 100 mg xgxr001  TRUE
#> 94  100 mg 100 mg xgxr001  TRUE
#> 95  100 mg 100 mg xgxr001  TRUE
#> 96  100 mg 100 mg xgxr001  TRUE
#> 97  100 mg 100 mg xgxr001  TRUE
#> 98  100 mg 100 mg xgxr001  TRUE
#> 99  100 mg 100 mg xgxr001  TRUE
#> 100 100 mg 100 mg xgxr001  TRUE
#> 101 100 mg 100 mg xgxr001  TRUE
#> 102 100 mg 100 mg xgxr001  TRUE
#> 103 100 mg 100 mg xgxr001  TRUE
#> 104 100 mg 100 mg xgxr001  TRUE
#> 105 100 mg 100 mg xgxr001  TRUE
#> 106 100 mg 100 mg xgxr001  TRUE
#> 107 100 mg 100 mg xgxr001  TRUE
#> 108 100 mg 100 mg xgxr001  TRUE
#> 109 100 mg 100 mg xgxr001  TRUE
#> 110 100 mg 100 mg xgxr001  TRUE
#> 111 100 mg 100 mg xgxr001  TRUE
#> 112 100 mg 100 mg xgxr001  TRUE
#> 113 100 mg 100 mg xgxr001  TRUE
#> 114 100 mg 100 mg xgxr001  TRUE
#> 115 100 mg 100 mg xgxr001  TRUE
#> 116 100 mg 100 mg xgxr001  TRUE
#> 117 100 mg 100 mg xgxr001  TRUE
#> 118 100 mg 100 mg xgxr001  TRUE
#> 119 100 mg 100 mg xgxr001  TRUE
#> 120 100 mg 100 mg xgxr001  TRUE
#> 121 300 mg 300 mg xgxr001  TRUE
#> 122 300 mg 300 mg xgxr001  TRUE
#> 123 300 mg 300 mg xgxr001  TRUE
#> 124 300 mg 300 mg xgxr001  TRUE
#> 125 300 mg 300 mg xgxr001  TRUE
#> 126 300 mg 300 mg xgxr001  TRUE
#> 127 300 mg 300 mg xgxr001  TRUE
#> 128 300 mg 300 mg xgxr001  TRUE
#> 129 300 mg 300 mg xgxr001  TRUE
#> 130 300 mg 300 mg xgxr001  TRUE
#> 131 300 mg 300 mg xgxr001  TRUE
#> 132 300 mg 300 mg xgxr001  TRUE
#> 133 300 mg 300 mg xgxr001  TRUE
#> 134 300 mg 300 mg xgxr001  TRUE
#> 135 300 mg 300 mg xgxr001  TRUE
#> 136 300 mg 300 mg xgxr001  TRUE
#> 137 300 mg 300 mg xgxr001  TRUE
#> 138 300 mg 300 mg xgxr001  TRUE
#> 139 300 mg 300 mg xgxr001  TRUE
#> 140 300 mg 300 mg xgxr001  TRUE
#> 141 300 mg 300 mg xgxr001  TRUE
#> 142 300 mg 300 mg xgxr001  TRUE
#> 143 300 mg 300 mg xgxr001  TRUE
#> 144 300 mg 300 mg xgxr001  TRUE
#> 145 300 mg 300 mg xgxr001  TRUE
#> 146 300 mg 300 mg xgxr001  TRUE
#> 147 300 mg 300 mg xgxr001  TRUE
#> 148 300 mg 300 mg xgxr001  TRUE
#> 149 300 mg 300 mg xgxr001  TRUE
#> 150 300 mg 300 mg xgxr001  TRUE
### Without an ID column we get non-varying columns
findCovs(dat)
#>   FLAG STUDY   TVKA  TVV2    TVCL    TVV3     TVQ      V3       Q BLQ CYCLE
#> 1    0     1 0.1812 0.042 0.72457 0.17848 2307400 0.17848 2307400   0     1
#>   PART PROFDAY TIMEUNIT   model nmout
#> 1    1       1    Hours xgxr001  TRUE