Extract columns that do not vary within variables in data

Extract columns that do not vary within variables in data

findCovs(data, cols.id = NULL, as.fun = NULL)

Arguments

data
data.frame in which to look for covariates

cols.id
covariates will be searched for in combinations of values in these columns. Often cols.id will be either empty or ID. But it can also be both say c("ID","DRUG") or c("ID","TRT").

as.fun
The default is to return a data.table if data is a data.table and return a data.frame in all other cases. Pass a function in as.fun to convert to something else. See ?runAsFun.

Value

a data set with one observation per combination of values of variables listed in cols.id.

See also

Examples

dat <- NMscanData(system.file("examples/nonmem/xgxr001.lst", package = "NMdata"))

#> In NMscanData(system.file("examples/nonmem/xgxr001.lst", package = "NMdata")) :
#> Input data columns will be appended to output data. However, column(s) were
#> identified as unique identifiers, present in both input and output data. If
#> this column or one of these columns is not modified by the Nonmem run,
#> consider using this in col.row for a robust merge of input and output data.
#> Candidate columns: ROW To skip this check, please specify either col.row
#> (recommended) or cbind.by.filters.

#> 

#> Warning: In NMscanData(system.file("examples/nonmem/xgxr001.lst", package = "NMdata")) :
#> input control stream
#> (C:/Users/delff/AppData/Local/Temp/RtmpuKKj5P/temp_libpath70c489a1541/NMdata/examples/nonmem/xgxr001.mod)
#> is newer than output control stream
#> (C:/Users/delff/AppData/Local/Temp/RtmpuKKj5P/temp_libpath70c489a1541/NMdata/examples/nonmem/xgxr001.lst)
#> Seems like model has been edited since last run. If data sections have been
#> edited, this can corrupt results.

#> Model:  xgxr001 
#> 
#> Tables, number of columns in tables, and their detail level:
#>            table used/total level
#>  xgxr001_res.txt      16/16   row
#>            input      22/22   row
#> 
#> Numbers of ID's and rows in data
#>       N From output tables From input data only
#>  N.rows                905                    0
#>   N.ids                150                    0
#> 
#> Distribution of rows on event types
#>  EVID From output tables
#>     0                755
#>     1                150

### very common use
findCovs(dat,cols.id="ID")

#>      ID FLAG STUDY    TVKA  TVV2    TVCL    TVV3     TVQ       KA       V2
#> 1    31    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 2    32    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 3    33    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 4    34    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 5    35    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.200980 0.044867
#> 6    36    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 7    37    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 8    38    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 9    39    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 10   40    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.198940 0.044708
#> 11   41    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.197620 0.044186
#> 12   42    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 13   43    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.210460 0.046150
#> 14   44    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 15   45    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 16   46    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 17   47    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.193710 0.044405
#> 18   48    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.193420 0.044558
#> 19   49    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 20   50    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.199420 0.044395
#> 21   51    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.232920 0.051853
#> 22   52    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 23   53    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 24   54    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.197710 0.045296
#> 25   55    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.196600 0.045040
#> 26   56    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 27   57    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 28   58    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 29   59    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.236010 0.052070
#> 30   60    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181160 0.042000
#> 31   61    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.226490 0.050277
#> 32   62    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.182800 0.042699
#> 33   63    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.157820 0.038191
#> 34   64    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.167120 0.040147
#> 35   65    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.175540 0.041752
#> 36   66    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.156170 0.038461
#> 37   67    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.154820 0.037750
#> 38   68    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.207850 0.045044
#> 39   69    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.200940 0.043803
#> 40   70    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.177070 0.041886
#> 41   71    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.197000 0.044794
#> 42   72    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.193850 0.049011
#> 43   73    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.192760 0.049967
#> 44   74    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.238040 0.053145
#> 45   75    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.224240 0.051677
#> 46   76    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.184790 0.042526
#> 47   77    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.243810 0.059397
#> 48   78    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.154970 0.038192
#> 49   79    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.183190 0.040635
#> 50   80    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.168150 0.040651
#> 51   81    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.306360 0.059155
#> 52   82    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.164380 0.039374
#> 53   83    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.217370 0.049527
#> 54   84    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.195560 0.042991
#> 55   85    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.242330 0.053778
#> 56   86    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.198640 0.046318
#> 57   87    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.202980 0.048004
#> 58   88    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.196660 0.046519
#> 59   89    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.180460 0.041720
#> 60   90    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.176070 0.041943
#> 61   91    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.304040 0.058857
#> 62   92    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.135840 0.041109
#> 63   93    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.196360 0.032508
#> 64   94    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.152420 0.035164
#> 65   95    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.160710 0.040626
#> 66   96    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.159820 0.038148
#> 67   97    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.164980 0.033916
#> 68   98    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.111950 0.063151
#> 69   99    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.191610 0.045673
#> 70  100    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.139770 0.032785
#> 71  101    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.218490 0.038332
#> 72  102    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.284460 0.044049
#> 73  103    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.111030 0.042973
#> 74  104    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.146970 0.037805
#> 75  105    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.311480 0.048335
#> 76  106    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.140450 0.035079
#> 77  107    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.192910 0.035892
#> 78  108    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.150810 0.036938
#> 79  109    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.122350 0.045355
#> 80  110    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.205140 0.036559
#> 81  111    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.298110 0.041774
#> 82  112    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.235970 0.043731
#> 83  113    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.200640 0.050461
#> 84  114    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.163250 0.044300
#> 85  115    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.191250 0.035263
#> 86  116    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.141670 0.033440
#> 87  117    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.107440 0.030178
#> 88  118    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.147590 0.031952
#> 89  119    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.196160 0.034625
#> 90  120    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.117460 0.032440
#> 91  121    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.220590 0.038723
#> 92  122    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.136500 0.038012
#> 93  123    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.160980 0.042487
#> 94  124    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.219600 0.033770
#> 95  125    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.153720 0.052751
#> 96  126    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.295100 0.042108
#> 97  127    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.075578 0.086706
#> 98  128    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.139970 0.039734
#> 99  129    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.204490 0.044726
#> 100 130    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.190000 0.042021
#> 101 131    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.084683 0.065548
#> 102 132    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.103500 0.069220
#> 103 133    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.205740 0.032194
#> 104 134    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.614680 0.036342
#> 105 135    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.302330 0.041101
#> 106 136    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.090507 0.041428
#> 107 137    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.331930 0.031867
#> 108 138    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.119460 0.037087
#> 109 139    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.091146 0.075929
#> 110 140    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.308080 0.040089
#> 111 141    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.087996 0.047212
#> 112 142    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.214520 0.028516
#> 113 143    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.110090 0.063183
#> 114 144    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.142500 0.058911
#> 115 145    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.221820 0.046501
#> 116 146    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.181220 0.051077
#> 117 147    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.167850 0.036629
#> 118 148    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.144050 0.048443
#> 119 149    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.330020 0.031505
#> 120 150    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.240580 0.026797
#> 121 151    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.074433 0.092395
#> 122 152    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.116810 0.051465
#> 123 153    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.422940 0.034489
#> 124 154    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.096618 0.051255
#> 125 155    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.151840 0.035464
#> 126 156    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.370540 0.022131
#> 127 157    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.170190 0.066922
#> 128 158    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.125960 0.056415
#> 129 159    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.133700 0.047473
#> 130 160    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.406330 0.020624
#> 131 161    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.144910 0.069496
#> 132 162    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.250630 0.020414
#> 133 163    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.192570 0.024115
#> 134 164    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.123500 0.061082
#> 135 165    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.401870 0.020528
#> 136 166    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.174450 0.038948
#> 137 167    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.282040 0.033692
#> 138 168    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.285040 0.040023
#> 139 169    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.116410 0.051393
#> 140 170    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.297140 0.021851
#> 141 171    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.091617 0.054356
#> 142 172    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.355320 0.022144
#> 143 173    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.324240 0.038398
#> 144 174    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.135420 0.057353
#> 145 175    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.312270 0.036866
#> 146 176    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.112360 0.058950
#> 147 177    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.134660 0.041848
#> 148 178    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.158520 0.051512
#> 149 179    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.148380 0.067208
#> 150 180    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.221310 0.022291
#>          CL      V3       Q BLQ CYCLE DOSE PART PROFDAY WEIGHTB   eff0 nmout
#> 1   0.72491 0.17857 8813700   0     1    3    1       1  87.031 56.461  TRUE
#> 2   0.72491 0.17857 8813700   0     1    3    1       1 100.620 45.096  TRUE
#> 3   0.72491 0.17857 8813700   0     1    3    1       1 112.990 49.189  TRUE
#> 4   0.72491 0.17857 8813700   0     1    3    1       1 144.130 56.239  TRUE
#> 5   0.72487 0.17857 8813700   0     1    3    1       1  80.384 58.419  TRUE
#> 6   0.72491 0.17857 8813700   0     1    3    1       1 120.860 58.163  TRUE
#> 7   0.72491 0.17857 8813700   0     1    3    1       1 138.440 42.122  TRUE
#> 8   0.72491 0.17857 8813700   0     1    3    1       1 123.350 59.180  TRUE
#> 9   0.72491 0.17857 8813700   0     1    3    1       1 137.080 47.222  TRUE
#> 10  0.72487 0.17857 8813700   0     1    3    1       1 134.090 49.498  TRUE
#> 11  0.72488 0.17857 8813700   0     1    3    1       1  80.018 41.362  TRUE
#> 12  0.72491 0.17857 8813700   0     1    3    1       1 142.900 54.624  TRUE
#> 13  0.72486 0.17857 8813700   0     1    3    1       1  97.637 51.154  TRUE
#> 14  0.72491 0.17857 8813700   0     1    3    1       1 105.310 54.303  TRUE
#> 15  0.72491 0.17857 8813700   0     1    3    1       1  83.345 59.970  TRUE
#> 16  0.72491 0.17857 8813700   0     1    3    1       1 148.540 40.533  TRUE
#> 17  0.72487 0.17857 8813700   0     1    3    1       1 121.060 42.542  TRUE
#> 18  0.72487 0.17857 8813700   0     1    3    1       1  97.761 51.290  TRUE
#> 19  0.72491 0.17857 8813700   0     1    3    1       1 144.240 45.758  TRUE
#> 20  0.72488 0.17857 8813700   0     1    3    1       1 111.400 44.106  TRUE
#> 21  0.72478 0.17857 8813700   0     1    3    1       1 144.870 42.832  TRUE
#> 22  0.72491 0.17857 8813700   0     1    3    1       1 136.980 48.142  TRUE
#> 23  0.72491 0.17857 8813700   0     1    3    1       1 124.950 52.037  TRUE
#> 24  0.72486 0.17857 8813700   0     1    3    1       1 139.260 54.637  TRUE
#> 25  0.72486 0.17857 8813700   0     1    3    1       1 127.350 57.822  TRUE
#> 26  0.72491 0.17857 8813700   0     1    3    1       1 136.260 48.192  TRUE
#> 27  0.72491 0.17857 8813700   0     1    3    1       1 139.740 45.235  TRUE
#> 28  0.72491 0.17857 8813700   0     1    3    1       1  90.185 51.918  TRUE
#> 29  0.72478 0.17857 8813700   0     1    3    1       1 142.720 54.874  TRUE
#> 30  0.72491 0.17857 8813700   0     1    3    1       1  91.702 57.376  TRUE
#> 31  0.72479 0.17857 8813700   0     1   10    1       1 133.040 58.744  TRUE
#> 32  0.72490 0.17857 8813700   0     1   10    1       1 104.290 40.081  TRUE
#> 33  0.72496 0.17857 8813700   0     1   10    1       1 107.650 49.956  TRUE
#> 34  0.72493 0.17857 8813700   0     1   10    1       1 141.420 56.587  TRUE
#> 35  0.72490 0.17857 8813700   0     1   10    1       1 113.970 50.127  TRUE
#> 36  0.72495 0.17857 8813700   0     1   10    1       1 112.190 49.280  TRUE
#> 37  0.72497 0.17857 8813700   0     1   10    1       1  87.380 48.628  TRUE
#> 38  0.72487 0.17857 8813700   0     1   10    1       1 126.680 47.701  TRUE
#> 39  0.72492 0.17857 8813700   0     1   10    1       1 110.810 44.421  TRUE
#> 40  0.72490 0.17857 8813700   0     1   10    1       1 142.900 50.431  TRUE
#> 41  0.72488 0.17857 8813700   0     1   10    1       1  88.158 58.193  TRUE
#> 42  0.72481 0.17857 8813700   0     1   10    1       1 115.390 59.961  TRUE
#> 43  0.72477 0.17857 8813700   0     1   10    1       1  91.806 41.836  TRUE
#> 44  0.72475 0.17857 8813700   0     1   10    1       1 111.080 42.848  TRUE
#> 45  0.72475 0.17857 8813700   0     1   10    1       1 120.870 56.245  TRUE
#> 46  0.72490 0.17857 8813700   0     1   10    1       1 102.040 47.031  TRUE
#> 47  0.72466 0.17857 8813700   0     1   10    1       1 143.750 52.876  TRUE
#> 48  0.72495 0.17857 8813700   0     1   10    1       1 148.760 46.962  TRUE
#> 49  0.72495 0.17857 8813700   0     1   10    1       1 128.350 43.855  TRUE
#> 50  0.72491 0.17857 8813700   0     1   10    1       1 143.310 57.924  TRUE
#> 51  0.72476 0.17857 8813700   0     1   10    1       1 145.680 48.497  TRUE
#> 52  0.72493 0.17857 8813700   0     1   10    1       1 125.550 48.890  TRUE
#> 53  0.72480 0.17857 8813700   0     1   10    1       1 124.870 41.692  TRUE
#> 54  0.72490 0.17857 8813700   0     1   10    1       1 127.300 47.241  TRUE
#> 55  0.72476 0.17857 8813700   0     1   10    1       1 139.600 55.084  TRUE
#> 56  0.72482 0.17857 8813700   0     1   10    1       1 147.170 42.582  TRUE
#> 57  0.72481 0.17857 8813700   0     1   10    1       1 112.100 59.316  TRUE
#> 58  0.72484 0.17857 8813700   0     1   10    1       1 122.970 49.694  TRUE
#> 59  0.72491 0.17857 8813700   0     1   10    1       1 101.600 40.517  TRUE
#> 60  0.72489 0.17857 8813700   0     1   10    1       1 126.330 46.436  TRUE
#> 61  0.72466 0.17857 8813700   0     1   30    1       1  91.716 58.817  TRUE
#> 62  0.72484 0.17857 8813700   0     1   30    1       1 101.980 41.972  TRUE
#> 63  0.72516 0.17857 8813700   0     1   30    1       1 131.860 57.696  TRUE
#> 64  0.72497 0.17857 8813700   0     1   30    1       1  84.754 42.902  TRUE
#> 65  0.72485 0.17857 8813700   0     1   30    1       1  83.878 41.517  TRUE
#> 66  0.72487 0.17857 8813700   0     1   30    1       1 121.110 57.137  TRUE
#> 67  0.72504 0.17857 8813700   0     1   30    1       1 133.690 55.847  TRUE
#> 68  0.72463 0.17857 8813700   0     1   30    1       1  80.206 55.302  TRUE
#> 69  0.72480 0.17857 8813700   0     1   30    1       1  96.745 49.018  TRUE
#> 70  0.72504 0.17857 8813700   0     1   30    1       1 109.040 46.163  TRUE
#> 71  0.72499 0.17857 8813700   0     1   30    1       1 142.040 53.020  TRUE
#> 72  0.72490 0.17857 8813700   0     1   30    1       1 111.700 53.996  TRUE
#> 73  0.72490 0.17857 8813700   0     1   30    1       1  92.732 50.635  TRUE
#> 74  0.72495 0.17857 8813700   0     1   30    1       1  83.444 58.264  TRUE
#> 75  0.72488 0.17857 8813700   0     1   30    1       1 121.690 49.004  TRUE
#> 76  0.72502 0.17857 8813700   0     1   30    1       1 133.300 53.440  TRUE
#> 77  0.72504 0.17857 8813700   0     1   30    1       1  84.755 49.630  TRUE
#> 78  0.72502 0.17857 8813700   0     1   30    1       1  95.846 46.434  TRUE
#> 79  0.72485 0.17857 8813700   0     1   30    1       1 113.810 57.663  TRUE
#> 80  0.72508 0.17857 8813700   0     1   30    1       1 116.780 58.513  TRUE
#> 81  0.72494 0.17857 8813700   0     1   30    1       1 138.870 58.687  TRUE
#> 82  0.72487 0.17857 8813700   0     1   30    1       1 114.000 50.922  TRUE
#> 83  0.72475 0.17857 8813700   0     1   30    1       1 114.260 49.663  TRUE
#> 84  0.72477 0.17857 8813700   0     1   30    1       1 105.090 49.476  TRUE
#> 85  0.72500 0.17857 8813700   0     1   30    1       1 129.360 50.889  TRUE
#> 86  0.72501 0.17857 8813700   0     1   30    1       1 109.160 42.200  TRUE
#> 87  0.72509 0.17857 8813700   0     1   30    1       1 116.960 45.204  TRUE
#> 88  0.72510 0.17857 8813700   0     1   30    1       1 119.330 40.657  TRUE
#> 89  0.72503 0.17857 8813700   0     1   30    1       1 130.040 57.173  TRUE
#> 90  0.72506 0.17857 8813700   0     1   30    1       1  95.384 47.696  TRUE
#> 91  0.72486 0.17857 8813700   0     1  100    1       1 136.770 47.184  TRUE
#> 92  0.72480 0.17857 8813700   0     1  100    1       1  94.357 49.776  TRUE
#> 93  0.72488 0.17857 8813700   0     1  100    1       1  81.491 57.013  TRUE
#> 94  0.72500 0.17857 8813700   0     1  100    1       1 115.370 51.224  TRUE
#> 95  0.72482 0.17857 8813700   0     1  100    1       1 148.110 41.539  TRUE
#> 96  0.72494 0.17857 8813700   0     1  100    1       1 148.230 48.961  TRUE
#> 97  0.72434 0.17857 8813700   0     1  100    1       1 133.160 42.732  TRUE
#> 98  0.72512 0.17857 8813700   0     1  100    1       1  82.140 49.704  TRUE
#> 99  0.72482 0.17857 8813700   0     1  100    1       1  88.087 46.169  TRUE
#> 100 0.72483 0.17857 8813700   0     1  100    1       1 142.170 49.395  TRUE
#> 101 0.72459 0.17857 8813700   0     1  100    1       1 117.810 57.872  TRUE
#> 102 0.72460 0.17857 8813700   0     1  100    1       1 132.710 42.860  TRUE
#> 103 0.72504 0.17857 8813700   0     1  100    1       1 114.950 40.940  TRUE
#> 104 0.72507 0.17857 8813700   0     1  100    1       1 145.100 54.023  TRUE
#> 105 0.72483 0.17857 8813700   0     1  100    1       1 117.550 55.676  TRUE
#> 106 0.72479 0.17857 8813700   0     1  100    1       1 117.820 45.134  TRUE
#> 107 0.72518 0.17857 8813700   0     1  100    1       1 133.540 45.305  TRUE
#> 108 0.72494 0.17857 8813700   0     1  100    1       1 119.450 49.917  TRUE
#> 109 0.72449 0.17857 8813700   0     1  100    1       1 149.550 43.930  TRUE
#> 110 0.72481 0.17857 8813700   0     1  100    1       1 112.980 48.596  TRUE
#> 111 0.72467 0.17857 8813700   0     1  100    1       1 130.200 44.755  TRUE
#> 112 0.72523 0.17857 8813700   0     1  100    1       1 147.080 50.895  TRUE
#> 113 0.72466 0.17857 8813700   0     1  100    1       1  96.265 43.053  TRUE
#> 114 0.72470 0.17857 8813700   0     1  100    1       1 132.650 51.302  TRUE
#> 115 0.72497 0.17857 8813700   0     1  100    1       1 118.270 50.193  TRUE
#> 116 0.72488 0.17857 8813700   0     1  100    1       1 114.600 56.089  TRUE
#> 117 0.72486 0.17857 8813700   0     1  100    1       1 141.930 54.312  TRUE
#> 118 0.72480 0.17857 8813700   0     1  100    1       1  99.604 44.034  TRUE
#> 119 0.72511 0.17857 8813700   0     1  100    1       1 111.110 55.695  TRUE
#> 120 0.72508 0.17857 8813700   0     1  100    1       1 111.820 55.035  TRUE
#> 121 0.72424 0.17857 8813700   0     1  300    1       1  88.319 53.035  TRUE
#> 122 0.72468 0.17857 8813700   0     1  300    1       1 121.320 47.391  TRUE
#> 123 0.72501 0.17857 8813700   0     1  300    1       1  93.475 56.259  TRUE
#> 124 0.72494 0.17857 8813700   0     1  300    1       1 147.680 52.824  TRUE
#> 125 0.72498 0.17857 8813700   0     1  300    1       1  96.953 58.300  TRUE
#> 126 0.72508 0.17857 8813700   0     1  300    1       1 140.810 46.160  TRUE
#> 127 0.72488 0.17857 8813700   0     1  300    1       1  94.526 53.894  TRUE
#> 128 0.72457 0.17857 8813700   0     1  300    1       1 146.820 54.044  TRUE
#> 129 0.72475 0.17857 8813700   0     1  300    1       1 145.530 55.699  TRUE
#> 130 0.72521 0.17857 8813700   0     1  300    1       1 127.400 50.455  TRUE
#> 131 0.72523 0.17857 8813700   0     1  300    1       1 101.580 55.400  TRUE
#> 132 0.72526 0.17857 8813700   0     1  300    1       1 102.620 59.111  TRUE
#> 133 0.72518 0.17857 8813700   0     1  300    1       1  85.221 41.015  TRUE
#> 134 0.72469 0.17857 8813700   0     1  300    1       1  84.974 54.064  TRUE
#> 135 0.72531 0.17857 8813700   0     1  300    1       1 124.230 53.337  TRUE
#> 136 0.72496 0.17857 8813700   0     1  300    1       1 101.180 59.792  TRUE
#> 137 0.72506 0.17857 8813700   0     1  300    1       1 119.050 45.097  TRUE
#> 138 0.72495 0.17857 8813700   0     1  300    1       1  95.155 46.504  TRUE
#> 139 0.72463 0.17857 8813700   0     1  300    1       1 124.980 54.155  TRUE
#> 140 0.72537 0.17857 8813700   0     1  300    1       1 102.150 51.862  TRUE
#> 141 0.72470 0.17857 8813700   0     1  300    1       1 136.230 47.295  TRUE
#> 142 0.72537 0.17857 8813700   0     1  300    1       1  80.888 49.242  TRUE
#> 143 0.72517 0.17857 8813700   0     1  300    1       1 132.040 53.646  TRUE
#> 144 0.72465 0.17857 8813700   0     1  300    1       1 118.690 59.691  TRUE
#> 145 0.72518 0.17857 8813700   0     1  300    1       1 111.830 53.775  TRUE
#> 146 0.72475 0.17857 8813700   0     1  300    1       1  95.780 54.516  TRUE
#> 147 0.72495 0.17857 8813700   0     1  300    1       1 135.530 51.307  TRUE
#> 148 0.72486 0.17857 8813700   0     1  300    1       1  99.496 51.160  TRUE
#> 149 0.72512 0.17857 8813700   0     1  300    1       1  89.308 55.116  TRUE
#> 150 0.72520 0.17857 8813700   0     1  300    1       1  98.103 57.835  TRUE
#>     TIMEUNIT TRTACT         flag trtact   model
#> 1      Hours   3 mg Analysis set   3 mg xgxr001
#> 2      Hours   3 mg Analysis set   3 mg xgxr001
#> 3      Hours   3 mg Analysis set   3 mg xgxr001
#> 4      Hours   3 mg Analysis set   3 mg xgxr001
#> 5      Hours   3 mg Analysis set   3 mg xgxr001
#> 6      Hours   3 mg Analysis set   3 mg xgxr001
#> 7      Hours   3 mg Analysis set   3 mg xgxr001
#> 8      Hours   3 mg Analysis set   3 mg xgxr001
#> 9      Hours   3 mg Analysis set   3 mg xgxr001
#> 10     Hours   3 mg Analysis set   3 mg xgxr001
#> 11     Hours   3 mg Analysis set   3 mg xgxr001
#> 12     Hours   3 mg Analysis set   3 mg xgxr001
#> 13     Hours   3 mg Analysis set   3 mg xgxr001
#> 14     Hours   3 mg Analysis set   3 mg xgxr001
#> 15     Hours   3 mg Analysis set   3 mg xgxr001
#> 16     Hours   3 mg Analysis set   3 mg xgxr001
#> 17     Hours   3 mg Analysis set   3 mg xgxr001
#> 18     Hours   3 mg Analysis set   3 mg xgxr001
#> 19     Hours   3 mg Analysis set   3 mg xgxr001
#> 20     Hours   3 mg Analysis set   3 mg xgxr001
#> 21     Hours   3 mg Analysis set   3 mg xgxr001
#> 22     Hours   3 mg Analysis set   3 mg xgxr001
#> 23     Hours   3 mg Analysis set   3 mg xgxr001
#> 24     Hours   3 mg Analysis set   3 mg xgxr001
#> 25     Hours   3 mg Analysis set   3 mg xgxr001
#> 26     Hours   3 mg Analysis set   3 mg xgxr001
#> 27     Hours   3 mg Analysis set   3 mg xgxr001
#> 28     Hours   3 mg Analysis set   3 mg xgxr001
#> 29     Hours   3 mg Analysis set   3 mg xgxr001
#> 30     Hours   3 mg Analysis set   3 mg xgxr001
#> 31     Hours  10 mg Analysis set  10 mg xgxr001
#> 32     Hours  10 mg Analysis set  10 mg xgxr001
#> 33     Hours  10 mg Analysis set  10 mg xgxr001
#> 34     Hours  10 mg Analysis set  10 mg xgxr001
#> 35     Hours  10 mg Analysis set  10 mg xgxr001
#> 36     Hours  10 mg Analysis set  10 mg xgxr001
#> 37     Hours  10 mg Analysis set  10 mg xgxr001
#> 38     Hours  10 mg Analysis set  10 mg xgxr001
#> 39     Hours  10 mg Analysis set  10 mg xgxr001
#> 40     Hours  10 mg Analysis set  10 mg xgxr001
#> 41     Hours  10 mg Analysis set  10 mg xgxr001
#> 42     Hours  10 mg Analysis set  10 mg xgxr001
#> 43     Hours  10 mg Analysis set  10 mg xgxr001
#> 44     Hours  10 mg Analysis set  10 mg xgxr001
#> 45     Hours  10 mg Analysis set  10 mg xgxr001
#> 46     Hours  10 mg Analysis set  10 mg xgxr001
#> 47     Hours  10 mg Analysis set  10 mg xgxr001
#> 48     Hours  10 mg Analysis set  10 mg xgxr001
#> 49     Hours  10 mg Analysis set  10 mg xgxr001
#> 50     Hours  10 mg Analysis set  10 mg xgxr001
#> 51     Hours  10 mg Analysis set  10 mg xgxr001
#> 52     Hours  10 mg Analysis set  10 mg xgxr001
#> 53     Hours  10 mg Analysis set  10 mg xgxr001
#> 54     Hours  10 mg Analysis set  10 mg xgxr001
#> 55     Hours  10 mg Analysis set  10 mg xgxr001
#> 56     Hours  10 mg Analysis set  10 mg xgxr001
#> 57     Hours  10 mg Analysis set  10 mg xgxr001
#> 58     Hours  10 mg Analysis set  10 mg xgxr001
#> 59     Hours  10 mg Analysis set  10 mg xgxr001
#> 60     Hours  10 mg Analysis set  10 mg xgxr001
#> 61     Hours  30 mg Analysis set  30 mg xgxr001
#> 62     Hours  30 mg Analysis set  30 mg xgxr001
#> 63     Hours  30 mg Analysis set  30 mg xgxr001
#> 64     Hours  30 mg Analysis set  30 mg xgxr001
#> 65     Hours  30 mg Analysis set  30 mg xgxr001
#> 66     Hours  30 mg Analysis set  30 mg xgxr001
#> 67     Hours  30 mg Analysis set  30 mg xgxr001
#> 68     Hours  30 mg Analysis set  30 mg xgxr001
#> 69     Hours  30 mg Analysis set  30 mg xgxr001
#> 70     Hours  30 mg Analysis set  30 mg xgxr001
#> 71     Hours  30 mg Analysis set  30 mg xgxr001
#> 72     Hours  30 mg Analysis set  30 mg xgxr001
#> 73     Hours  30 mg Analysis set  30 mg xgxr001
#> 74     Hours  30 mg Analysis set  30 mg xgxr001
#> 75     Hours  30 mg Analysis set  30 mg xgxr001
#> 76     Hours  30 mg Analysis set  30 mg xgxr001
#> 77     Hours  30 mg Analysis set  30 mg xgxr001
#> 78     Hours  30 mg Analysis set  30 mg xgxr001
#> 79     Hours  30 mg Analysis set  30 mg xgxr001
#> 80     Hours  30 mg Analysis set  30 mg xgxr001
#> 81     Hours  30 mg Analysis set  30 mg xgxr001
#> 82     Hours  30 mg Analysis set  30 mg xgxr001
#> 83     Hours  30 mg Analysis set  30 mg xgxr001
#> 84     Hours  30 mg Analysis set  30 mg xgxr001
#> 85     Hours  30 mg Analysis set  30 mg xgxr001
#> 86     Hours  30 mg Analysis set  30 mg xgxr001
#> 87     Hours  30 mg Analysis set  30 mg xgxr001
#> 88     Hours  30 mg Analysis set  30 mg xgxr001
#> 89     Hours  30 mg Analysis set  30 mg xgxr001
#> 90     Hours  30 mg Analysis set  30 mg xgxr001
#> 91     Hours 100 mg Analysis set 100 mg xgxr001
#> 92     Hours 100 mg Analysis set 100 mg xgxr001
#> 93     Hours 100 mg Analysis set 100 mg xgxr001
#> 94     Hours 100 mg Analysis set 100 mg xgxr001
#> 95     Hours 100 mg Analysis set 100 mg xgxr001
#> 96     Hours 100 mg Analysis set 100 mg xgxr001
#> 97     Hours 100 mg Analysis set 100 mg xgxr001
#> 98     Hours 100 mg Analysis set 100 mg xgxr001
#> 99     Hours 100 mg Analysis set 100 mg xgxr001
#> 100    Hours 100 mg Analysis set 100 mg xgxr001
#> 101    Hours 100 mg Analysis set 100 mg xgxr001
#> 102    Hours 100 mg Analysis set 100 mg xgxr001
#> 103    Hours 100 mg Analysis set 100 mg xgxr001
#> 104    Hours 100 mg Analysis set 100 mg xgxr001
#> 105    Hours 100 mg Analysis set 100 mg xgxr001
#> 106    Hours 100 mg Analysis set 100 mg xgxr001
#> 107    Hours 100 mg Analysis set 100 mg xgxr001
#> 108    Hours 100 mg Analysis set 100 mg xgxr001
#> 109    Hours 100 mg Analysis set 100 mg xgxr001
#> 110    Hours 100 mg Analysis set 100 mg xgxr001
#> 111    Hours 100 mg Analysis set 100 mg xgxr001
#> 112    Hours 100 mg Analysis set 100 mg xgxr001
#> 113    Hours 100 mg Analysis set 100 mg xgxr001
#> 114    Hours 100 mg Analysis set 100 mg xgxr001
#> 115    Hours 100 mg Analysis set 100 mg xgxr001
#> 116    Hours 100 mg Analysis set 100 mg xgxr001
#> 117    Hours 100 mg Analysis set 100 mg xgxr001
#> 118    Hours 100 mg Analysis set 100 mg xgxr001
#> 119    Hours 100 mg Analysis set 100 mg xgxr001
#> 120    Hours 100 mg Analysis set 100 mg xgxr001
#> 121    Hours 300 mg Analysis set 300 mg xgxr001
#> 122    Hours 300 mg Analysis set 300 mg xgxr001
#> 123    Hours 300 mg Analysis set 300 mg xgxr001
#> 124    Hours 300 mg Analysis set 300 mg xgxr001
#> 125    Hours 300 mg Analysis set 300 mg xgxr001
#> 126    Hours 300 mg Analysis set 300 mg xgxr001
#> 127    Hours 300 mg Analysis set 300 mg xgxr001
#> 128    Hours 300 mg Analysis set 300 mg xgxr001
#> 129    Hours 300 mg Analysis set 300 mg xgxr001
#> 130    Hours 300 mg Analysis set 300 mg xgxr001
#> 131    Hours 300 mg Analysis set 300 mg xgxr001
#> 132    Hours 300 mg Analysis set 300 mg xgxr001
#> 133    Hours 300 mg Analysis set 300 mg xgxr001
#> 134    Hours 300 mg Analysis set 300 mg xgxr001
#> 135    Hours 300 mg Analysis set 300 mg xgxr001
#> 136    Hours 300 mg Analysis set 300 mg xgxr001
#> 137    Hours 300 mg Analysis set 300 mg xgxr001
#> 138    Hours 300 mg Analysis set 300 mg xgxr001
#> 139    Hours 300 mg Analysis set 300 mg xgxr001
#> 140    Hours 300 mg Analysis set 300 mg xgxr001
#> 141    Hours 300 mg Analysis set 300 mg xgxr001
#> 142    Hours 300 mg Analysis set 300 mg xgxr001
#> 143    Hours 300 mg Analysis set 300 mg xgxr001
#> 144    Hours 300 mg Analysis set 300 mg xgxr001
#> 145    Hours 300 mg Analysis set 300 mg xgxr001
#> 146    Hours 300 mg Analysis set 300 mg xgxr001
#> 147    Hours 300 mg Analysis set 300 mg xgxr001
#> 148    Hours 300 mg Analysis set 300 mg xgxr001
#> 149    Hours 300 mg Analysis set 300 mg xgxr001
#> 150    Hours 300 mg Analysis set 300 mg xgxr001

### Without an ID column we get non-varying columns
findCovs(dat)

#>   FLAG STUDY    TVKA  TVV2    TVCL    TVV3     TVQ      V3       Q BLQ CYCLE
#> 1    0     1 0.18116 0.042 0.72491 0.17857 8813700 0.17857 8813700   0     1
#>   PART PROFDAY nmout TIMEUNIT         flag   model
#> 1    1       1  TRUE    Hours Analysis set xgxr001