Assign exclusion flags to a dataset based on specified table

The aim with this function is to take a (say PK) dataset and a pre-specified table of flags, assign the flags automatically.

flagsAssign(
  data,
  tab.flags,
  return.all = F,
  col.id = "ID",
  col.dv = "DV",
  as.fun = NULL
)

Arguments

data
The dataset to assign flags to.

tab.flags
A data.frame containing at least these named columns: FLAG, flag, condition. Condition is disregarded for FLAG==0. FLAG must be numeric and non-negative, flag and condition are characters.

return.all
If TRUE, both the edited dataset and the table of flags are returned. If FALSE (default) only the edited dataset is returned.

col.id
The name of the subject ID column. Default is "ID".

col.dv
The name of the data value column. Default is "DV".

as.fun
The default is to return data in data.tables. Pass a function in as.fun to convert to something else. If data.frames are wanted, use as.fun=as.data.frame.

Value

The dataset with flags added. See parameter flags.return as well.

See also